% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Movements-methods.R
\docType{methods}
\name{accession}
\alias{accession}
\alias{accession,Movements-method}
\title{Calculate accession from a movements account.}
\usage{
accession(object, births = TRUE)

\S4method{accession}{Movements}(object, births = TRUE)
}
\arguments{
\item{object}{A \code{\linkS4class{Movements}} account.}

\item{births}{If \code{TRUE} (the default), births are included in the
answer; if \code{FALSE}, they are not.}
}
\value{
An object of class \code{\linkS4class{Counts}}.
}
\description{
Calculations accession from a movements account with an age dimension.
Accession is the number of people reaching a given age during a given
period.  For instance, accession to age 65 during 2019 is the number of
people turning 65 during 2019.
}
\details{
Births can be considered as accession to age 0.  By default,
the return value to \code{accession} includess age 0.
If the account does not in fact include births, then accession
to age 0 is set to 0.
}
\examples{
population <- Counts(array(c(10, 15, 13, 16),
                           dim = c(2, 2),
                           dimnames = list(age = c("0-29", "30+"),
                                           time = c(1970, 2000))))
births <- Counts(array(13,
                       dim = c(1, 1),
                       dimnames = list(age = "30+",
                                       time = "1971-2000")))
deaths <- Counts(array(c(0, 9),
                       dim = c(2, 1),
                       dimnames = list(age = c("0-29", "30+"),
                                       time = c("1971-2000"))))
account <- Movements(population = population,
                     births = births,
                     exits = list(deaths = deaths))

accession(account)
accession(account, births = FALSE)
}

