% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{addDest}
\alias{addDest}
\alias{addDest,Counts-method}
\title{Add a "destination" dimension to an object.}
\usage{
addDest(object, base)

\S4method{addDest}{Counts}(object, base)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}

\item{base}{The name of an existing dimension.  This dimension
must have \code{\link{dimtype}} \code{"state"} or \code{"sex"}.}
}
\value{
A modified version of \code{object}
}
\description{
Add a dimension with \code{\link{dimtype}} \code{"destination"}
to an object of class \code{\linkS4class{DemographicArray}}.
This is useful mainly when constructing an exposure measure
to use when measuring or modelling origin-destination
rates.
}
\details{
The new "destination" dimension uses the same categories as one of the
existing dimensions, which becomes the "origin" dimension.  The original
data is replicated once for each category within the destination
dimensions.  This means that in a model of origin-destination
rates, each destination has the same exposures.
}
\examples{
population <- CountsOne(10:12, labels = c("A", "B", "C"), name = "region")
exposure <- addDest(population, base = "region")
exposure
}
\seealso{
To add a dimension that does not have dimtype \code{"destination"},
use function \code{\link{addDimension}}.
}
