% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\docType{methods}
\name{addDimension}
\alias{addDimension}
\alias{addDimension,Counts,ANY,ANY,ANY,ANY,ANY,missing-method}
\alias{addDimension,Values,ANY,ANY,ANY,ANY,ANY,ANY-method}
\title{Add one or more dimensions to a demographic array.}
\usage{
addDimension(object, name, labels, after = length(dim(object)),
  dimtype = NULL, dimscale = NULL, scale = 1L, ...)

\S4method{addDimension}{Counts,ANY,ANY,ANY,ANY,ANY,missing}(object, name,
  labels, after = length(dim(object)), dimtype = NULL, dimscale = NULL)

\S4method{addDimension}{Values,ANY,ANY,ANY,ANY,ANY,ANY}(object, name, labels,
  after = length(dim(object)), dimtype = NULL, dimscale = NULL,
  scale = 1L)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}}

\item{name}{Character vector with name of dimension (or names of dimensions) to be added.}

\item{labels}{Character vector with labels to be used by new dimension - or,
if several dimensions are being added, and each dimension has different labels,
a list of character vectors.}

\item{after}{Name or index of a dimension of \code{object}.  The new dimension
will be added after this dimension.}

\item{dimtype}{Optional character with dimtype of new dimension
(or dimtypes of new dimensions.)}

\item{dimscale}{Optional character with dimscale of new dimension (or
dimscales of new dimensions.)}

\item{scale}{Optional numeric vector of scaling factors - or, if several dimensions
are being added, and each dimension is scaled differently, a list of numeric vectors.
Can only be used if \code{object} has class \code{\linkS4class{Values}}.}

\item{\dots}{Not currently used.}
}
\value{
Modified version of \code{object}
}
\description{
Add one or more dimensions to an object of class \code{\linkS4class{DemographicArray}}
}
\details{
If \code{object} has class \code{\linkS4class{Counts}}, then the new dimension(s)
must have length 1.  If \code{object} has class \code{\linkS4class{Values}},
then the new dimensions can be longer than 1, in which case the \code{scale}
argument becomes relevant.

\code{labels}, \code{scale}, \code{dimtype}, and \code{dimscale} are
all recycled where appropriate.  For instance, if \code{scale} is shorter than
\code{labels}, then \code{scales} is recycled until it matches the
length of \code{labels}.  Similarly, if \code{names} has length 2
but \code{labels} is a single vector, then the same \code{labels}
vector is used for both dimensions.
}
\examples{
library(demdata)
deaths <- Values(VADeaths2)
deaths
## assume that Democrats and Republicans have the same death rates
addDimension(deaths,
            name = "party",
            labels = c("Democrat", "Republican"))
## assume that Democrats have slightly higher death rates
addDimension(deaths,
            name = "party",
            labels = c("Democrat", "Republican"),
            scale = c(1.05, 0.95))
## put the new dimension in a different place
addDimension(deaths,
            name = "party",
            labels = c("Democrate", "Republican"),
            after = "age")

income <- Values(array(c(5000, 10000),
                      dim = 2,
                      dimnames = list(age = c("20-34", "35-64"))))
income
## add two dimensions
addDimension(income,
            name = c("residence_orig", "residence_dest"),
            labels = c("Urban", "Rural"))
## now assume that being in an urban area is associated with
## higher income - and note how the scales multiply
addDimension(income,
            name = c("residence_orig", "residence_dest"),
            labels = c("Urban", "Rural"),
            scale = c(1.1, 0.9))
## now with income depending only on origin
addDimension(income,
            name = c("residence_orig", "residence_dest"),
            labels = c("Urban", "Rural"),
            scale = list(c(1.1, 0.9), 1))
}
\seealso{
\code{\link{collapseDimension}}
}
