% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{addPair}
\alias{addPair}
\alias{addPair,Counts-method}
\title{Add a "destination" or "child" dimension to an object.}
\usage{
addPair(object, base, dimtype = c("destination", "child"))

\S4method{addPair}{Counts}(object, base, dimtype = c("destination", "child"))
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}

\item{base}{The name of an existing dimension.  This dimension
must have \code{\link{dimtype}} \code{"state"} or \code{"sex"}.}

\item{dimtype}{Either "destination" (the default) or "child".}
}
\value{
A modified version of \code{object}
}
\description{
Add a dimension with \code{\link{dimtype}} \code{"destination"}
or \code{"child"} to an object of class
\code{\linkS4class{DemographicArray}}.
This is useful mainly when constructing an exposure measure
to use when measuring or modelling origin-destination
or parent-child rates.
}
\details{
The new dimension uses the same categories as one of the
existing dimensions, which becomes the "origin" or "parent" dimension.
The original data is replicated once for each category
within the new dimension.  This means that in a model of origin-destination
rates, each destination has the same exposures, and in a model
or parent-child rates each child category has the same exposures.
}
\examples{
population <- CountsOne(10:12, labels = c("A", "B", "C"), name = "region")
addPair(population, base = "region")
addPair(population, base = "region", dimtype = "child")
}
\seealso{
To add a dimension that does not have dimtype \code{"destination"},
or \code{"child"}, use function \code{\link{addDimension}}.
}
