% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R, R/MetaData-methods.R
\docType{methods}
\name{ageMinMax}
\alias{ageMax}
\alias{ageMax,DemographicArray-method}
\alias{ageMax,MetaData-method}
\alias{ageMax<-}
\alias{ageMax<-,DemographicArray-method}
\alias{ageMax<-,MetaData-method}
\alias{ageMin}
\alias{ageMin,DemographicArray-method}
\alias{ageMin,MetaData-method}
\alias{ageMin<-}
\alias{ageMin<-,DemographicArray-method}
\alias{ageMin<-,MetaData-method}
\alias{ageMinMax}
\title{Get or set minimum and maximum ages.}
\usage{
ageMax(object)

ageMax(object) <- value

ageMin(object)

ageMin(object) <- value

\S4method{ageMax}{DemographicArray}(object)

\S4method{ageMax}{DemographicArray}(object) <- value

\S4method{ageMin}{DemographicArray}(object)

\S4method{ageMin}{DemographicArray}(object) <- value

\S4method{ageMax}{MetaData}(object)

\S4method{ageMax}{MetaData}(object) <- value

\S4method{ageMin}{MetaData}(object)

\S4method{ageMin}{MetaData}(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}
or \code{\linkS4class{DemographicAccount}}.}

\item{value}{A number. Can be \code{Inf}.}
}
\value{
The extraction functions return numbers, and the replacement
functions return a modified version of \code{object}.
}
\description{
Extract minimum or maximum ages from the \code{\link{dimscales}} of an object.
If the age dimension uses \code{dimscale} \code{"Intervals"}, then the
replacement version of \code{minAge} can be used to change the lower limit
of the first age group, and the replacement version of \code{maxAge} can be
used to change upper limit of the last age group.
}
\examples{
library(demdata)
popn <- Values(VADeaths2)
popn
ageMin(popn)
ageMax(popn)
ageMin(popn) <- 45
popn
ageMax(popn) <- Inf
popn
}
\seealso{
\code{\link{ageTimeStep}} and \code{\link{hasRegularAgeTime}} also
provide information about age groups.
}

