% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R,
%   R/MetaData-methods.R
\name{ageTimeStep}
\alias{ageTimeStep}
\alias{ageTimeStep,DemographicArray-method}
\alias{ageTimeStep,MetaData-method}
\title{Get the age-time step used by a demographic array.}
\usage{
ageTimeStep(object)

\S4method{ageTimeStep}{DemographicArray}(object)

\S4method{ageTimeStep}{MetaData}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}
}
\value{
A number (i.e. a numeric vector of length 1).
}
\description{
If a \code{object} has a \code{\link[=hasRegularAgeTime]{regular}} age-time plan,
then \code{ageTimeStep} returns the length of the age-time steps used;
otherwise it raises an error.
}
\examples{
library(demdata)
x <- Counts(VAPopn)
x
hasRegularAgeTime(x)
}
\seealso{
\code{\link{hasRegularAgeTime}}, \code{\linkS4class{DemographicArray}},
\code{\link{dimtypes}}, \code{\link{dimscales}}
}
