% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{alignPair}
\alias{alignPair}
\alias{alignPair,DemographicArray-method}
\title{Align categories dimensions that come in pairs.}
\usage{
alignPair(object, base = NULL, omitted = ifelse(methods::is(object,
  "Counts"), 0L, NA))

\S4method{alignPair}{DemographicArray}(object, base = NULL,
  omitted = ifelse(methods::is(object, "Counts"), 0L, NA))
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{base}{Character vector giving the "base names" for the dimension pairs.
If \code{base} is omitted, the base names from all dimensions with
\code{\link{dimtype}} \code{"origin"}, \code{"destination"}, \code{"parent"},
or \code{"child"} are used.}

\item{omitted}{Value to be used for omitted origin-destination or parent-child
combinations.  Defaults to \code{0} when \code{object} has class
\code{\linkS4class{Counts}} and \code{NA} otherwise.}
}
\value{
An object with the same class as \code{object}.
}
\description{
Given an object of class \code{\linkS4class{DemographicArray}}, make dimensions with
\code{\link{dimtype}} \code{"origin"} and \code{"destination"} or \code{"parent"}
and \code{"child"} consistent.  This is done by permuting and inserting rows.
}
\details{
The "base name" of a dimension with \code{\link{dimtype}} \code{"origin"},
\code{"destination"}, \code{"parent"}, or \code{"child"} is the name minus
the \code{"_orig"}, \code{"_dest"}, \code{"_parent"}, or \code{"_child"}
suffix.  For instance, the base name of \code{"region_orig"} and
\code{"region_dest"} is \code{"region"}.
}
\examples{
x <- Counts(array(1:4,
                 dim = c(2, 2),
                 dimnames = list(region_orig = c("a", "b"),
                 region_dest = c("c", "b"))))
x
alignPair(x, base = "region")
alignPair(x)
alignPair(x, omitted = NA)

y <- Values(array(1:8,
                  dim = c(2, 2, 1, 2),
                  dimnames = list(region_orig = c("a", "b"),
                                  region_dest = c("b", "a"),
                                  ethnicity_parent = "q",
                                  ethnicity_child = c("q", "r"))))
y
alignPair(y, base = "region")
alignPair(y, base = "ethnicity")
alignPair(y)
}
\seealso{
Function \code{\link{pairAligned}} tests whether a pair of dimensions
use the same categories in the same order.
}
