% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,Counts-method}
\alias{as.data.frame,Values-method}
\title{Convert a DemographicArray Object to a Data Frame}
\usage{
\S4method{as.data.frame}{Counts}(x, row.names = NULL, optional = FALSE,
  stringsAsFactors = TRUE, responseName = "count",
  direction = c("wide", "long"), midpoints = FALSE, ...)

\S4method{as.data.frame}{Values}(x, row.names = NULL, optional = FALSE,
  stringsAsFactors = TRUE, responseName = "value",
  direction = c("wide", "long"), midpoints = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{row.names}{'NULL' or a character vector giving the row names for the
data frame.  See \code{\link[base]{as.data.frame}}.}

\item{optional}{Logical.  See \code{\link[base]{as.data.frame}}.}

\item{stringsAsFactors}{Logical.  If \code{TRUE} (the default) factors are
used to represent qualitative variables.}

\item{responseName}{Name of column with numeric values when \code{direction}
is \code{"long"}.  Defaults to \code{"count"} if \code{object} is
\code{\linkS4class{Counts}} and \code{"value"} if \code{object} is
\code{\linkS4class{Values}}.}

\item{direction}{Logical.  Governs whether data are converted to
\code{"wide"} (the default) or \code{"long"} format.}

\item{midpoints}{\code{FALSE} (the default), \code{TRUE}, or names of
dimensions with \code{\link{dimscale}} \code{"Intervals"}.}

\item{\dots}{Additional arguments to be passed to or from methods.}
}
\value{
A data frame.
}
\description{
Convert a \code{"\linkS4class{DemographicArray}"} object to a data frame.  The
data frame can be a "wide" or "long" format.
}
\details{
If \code{direction} is \code{"wide"}, then \code{object} is flattened into a
matrix before conversion to a data frame, which is what the method for
ordinary arrays does.  If \code{direction} is \code{"long"}, then
\code{object} is put into "long" format, with a column to represent each
dimension, plus a final column holding the data.  The final column is called
\code{"count"} if \code{object} has class \code{\linkS4class{Counts}} and
\code{"value"} if \code{object} has class \code{\linkS4class{Counts}}.

If \code{midpoints} is \code{FALSE}, interval labels are converted into
factors.  If \code{midpoints} is \code{TRUE}, interval labels are replaced
by numbers giving their midpoints.  If names of individual dimensions are
supplied, then only the named dimensions are converted.
}
\examples{
library(demdata)
popn <- Counts(VAPopn)
as.data.frame(popn)
as.data.frame(popn, direction = "long")

x <- Values(array(1:4,
                  dim = c(2, 2),
                  dimnames = list(age = c("0-39", "40+"),
                      period = c("2001-2010", "2011-2020"))))
as.data.frame(x, direction = "long")
as.data.frame(x, direction = "long", midpoints = TRUE)
as.data.frame(x, direction = "long", midpoints = "period")
}
\seealso{
\code{\link[base]{as.data.frame}},
\code{\link[base]{as.data.frame.table}}, \code{\link[base]{data.frame}},
\code{\link{midpoints}}
}
