% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/CountsWithSubtotals-generator.R
\docType{methods}
\name{attachSubtotals}
\alias{attachSubtotals}
\alias{attachSubtotals,Counts,Counts-method}
\alias{attachSubtotals,Counts,Values-method}
\alias{attachSubtotals,Counts,numeric-method}
\title{Attach a set of known subtotals to a Counts object.}
\usage{
attachSubtotals(object, subtotals, concordances = list())

\S4method{attachSubtotals}{Counts,Counts}(object, subtotals,
  concordances = list())

\S4method{attachSubtotals}{Counts,Values}(object, subtotals,
  concordances = list())

\S4method{attachSubtotals}{Counts,numeric}(object, subtotals,
  concordances = list())
}
\arguments{
\item{object}{Object of class \code{\linkS4class{Counts}}, with missing
values.  All non-missing values must be non-negative integers.}

\item{subtotals}{Object of class \code{\linkS4class{Counts}}, or a single
number.  Missing values are not permitted.  All values must be non-negative
integers.}

\item{concordances}{A named list of
\code{\link[classconc:ManyToOne-class]{ManyToOne}} concordances.}
}
\value{
Object of class \code{\linkS4class{CountsWithSubtotals}}.
}
\description{
Attach a subtotal or set of subtotals to an object of class
\code{\linkS4class{Counts}} with missing values.
}
\details{
All values in \code{subtotals} and all non-missing values in \code{object}
must be non-negative integers.

Neither \code{object} nor \code{subtotals} may have \code{\link{dimtype}}
\code{"iteration"} or \code{"quantile"}.

Concordances supplied through the \code{concordances} argument are used
to map values from \code{object} on to \code{subtotals}.  The concordances
are supplied as a named list, with the names specifying which dimension
of \code{object} each concordances should be used with.  See below
for an example.
}
\examples{
## 'subtotals' have class "Counts"
popn <- demdata::VAPopn
popn <- Counts(popn)
popn <- extrapolate(popn, along = "age", labels = "45-49", type = "missing")
subtotals <- Counts(array(c(65000, 64000),
                          dim = 1:2,
			  dimnames = list(age = "45-49", sex = c("Male", "Female"))))
attachSubtotals(popn, subtotals = subtotals)

## 'subtotals' single number
attachSubtotals(popn, subtotals = 500000)

## 'concordances' argument
x <- CountsOne(c(1, 4, NA, NA, 7),
               labels = c("a", "b", "c", "d", "e"),
               name = "region")
subtotals <- CountsOne(c(10, 11),
                       labels = c("U", "V"),
                       name = "region")
conc <- data.frame(from = c("a", "b", "c", "d", "e"),
                   to = c("U", "U", "U", "V", "V"))
conc <- classconc::Concordance(conc)
attachSubtotals(x,
                subtotals = subtotals,
                concordances = list(region = conc))
}
\seealso{
\code{\link{impute}}, \code{\linkS4class{CountsWithSubtotals}}
}
