% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{chain}
\alias{chain}
\title{Chain concordances.}
\usage{
chain(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{Concordance}}.}
}
\value{
An object of class \code{\linkS4class{Concordance}}.
}
\description{
Combine two or more concordances to form a new concordance via a
\code{\link[base]{merge}}-like operation.  For instance, if \code{x} is a
concordance between classification \code{c1} and classification \code{c2},
and \code{y} is a concordance between \code{c2} and \code{c3}, then
\code{chain(x, y)} creates a concordance between \code{c1} and \code{c3}.
}
\details{
Two concordances that are to be chained together must have a classification
in common.  In addition, the two versions of the common classification must
have the same set of values, though not necessarily in the same order.
Concordances are merged from left to right.
}
\examples{
x <- cbind(c1 = c("a", "b", "c"), c2 = c("e", "f", "f"))
x <- Concordance(x)
y <- cbind(c2 = c("e", "f"), c3 = c("i", "j"))
y <- Concordance(y)
z <- cbind(c3 = c("i", "j"), c4 = c("m", "l"))
z <- Concordance(z)
x
y
z
chain(x, y)
chain(x, y, z)

x <- cbind(c1 = c("a", "b"), c2 = c("e", "f"))
x <- Concordance(x)
y <- cbind(c3 = c("i", "j"), c2 = c("e", "f"))
y <- Concordance(y)
z <- cbind(c3 = c("i", "j", "k"), c2 = c("e", "e", "f"))
z <- Concordance(z)
x
y
z
## The order of the columns is irrelevant in a "OneToOne" object
chain(x, y)
## ...but not a "ManyToOne" object
\dontrun{chain(x, z)}
}
\seealso{
\code{splice} combines concordances via a
\code{\link[base]{rbind}}-like operation.
}
