% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{checkAndTidyDateAndDOB}
\alias{checkAndTidyDateAndDOB}
\title{dateOfBirthToAgeCompleted - from date to years - (dateOfBirth, dateObs, completed)
dateOfBirthToAgeExact - from date to years - (dateOfBirth, dateObs, completed)
ageCompletedToAgeGroup - from years to multiyear age group - (age, breaks)
ageExactToMonth - from exact years to months - (age, min, max)
ageExactToQuarter - from exact years to quarters - (age, min, max)
dateToYear - from date to years - (date, month, day, labelStart) 
yearToPeriod - from years to multiyear age group - (year, breaks, labelStart)
dateToMonth - from date to month - (date)
dateToQuarter - from date to quarters - (date)
dateOfBirthToTriangle - (dateOfBirth, dateObs, month, day, width)
dateOfBirthToTriangleMonth - (dateOfBirth, dateObs)
dateOfBirthToTriangleQuarter - (dateOfBirth, dateObs)}
\usage{
checkAndTidyDateAndDOB(date, dob)
}
\description{
dateOfBirthToAgeCompleted - from date to years - (dateOfBirth, dateObs, completed)
dateOfBirthToAgeExact - from date to years - (dateOfBirth, dateObs, completed)
ageCompletedToAgeGroup - from years to multiyear age group - (age, breaks)
ageExactToMonth - from exact years to months - (age, min, max)
ageExactToQuarter - from exact years to quarters - (age, min, max)
dateToYear - from date to years - (date, month, day, labelStart) 
yearToPeriod - from years to multiyear age group - (year, breaks, labelStart)
dateToMonth - from date to month - (date)
dateToQuarter - from date to quarters - (date)
dateOfBirthToTriangle - (dateOfBirth, dateObs, month, day, width)
dateOfBirthToTriangleMonth - (dateOfBirth, dateObs)
dateOfBirthToTriangleQuarter - (dateOfBirth, dateObs)
}
