% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Concordance-methods.R
\docType{methods}
\name{classifications}
\alias{classifications}
\alias{classificationFrom}
\alias{classificationFrom<-}
\alias{classifications}
\alias{classifications<-}
\alias{classificationTo}
\alias{classificationTo<-}
\alias{classificationFrom,ManyToOne-method}
\alias{classifications,Concordance-method}
\alias{classifications<-,Concordance-method}
\alias{classificationTo,ManyToOne-method}
\title{Names of classifications used by concordances.}
\usage{
classificationFrom(object)

classificationFrom(object) <- value

classifications(object)

classifications(object) <- value

classificationTo(object)

classificationTo(object) <- value

\S4method{classificationFrom}{ManyToOne}(object)

\S4method{classifications}{Concordance}(object)

\S4method{classifications}{Concordance}(object) <- value

\S4method{classificationTo}{ManyToOne}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{Concordance}}.}

\item{value}{A character vector.}
}
\value{
A character vector, or in the case of the replacement function,
a concordance with new classification names.
}
\description{
Get and set names of classifications used by objects of class
\code{\link{Concordance}}.
}
\details{
\code{classificationFrom} and \code{classificationTo} can only be used with
concordances of class \code{\linkS4class{ManyToOne}}, since the idea of a
'from' classification or 'to' classification is only defined for a
many-to-one concordance. \code{classifications} can be used with concordances
of class \code{\linkS4class{ManyToOne}} or \code{\linkS4class{OneToOne}}.

\code{value} has length 2 with function \code{classifications} and length 1
with functions \code{classificationFrom} or \code{classificationTo}.
}
\examples{
x <- cbind(c1 = c("a", "b", "c"), c2 = c("x", "y", "x"))
x <- Concordance(x)
x
classifications(x)
classifications(x)[1] <- "classif1"
classifications(x)
classificationFrom(x)
classificationTo(x)
}
\seealso{
The codes from a concordance can be extracted using
\code{\link{codes}}.
}
