% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{cleanAgeGroup}
\alias{cleanAgeGroup}
\alias{cleanAgeGroupConc}
\title{Clean up age group labels}
\usage{
cleanAgeGroup(age)

cleanAgeGroupConc(age)
}
\arguments{
\item{age}{A vector of age labels.}
}
\value{
A character vector of reformatted age labels.
}
\description{
Reformat a vector of age-group labels, so that it matches the format
expected by functions such as \code{\link{Counts}} and
\code{\link{Values}}.
}
\details{
\code{cleanAgeGroup} strips off spaces or other symbols,
converts text to numbers, and otherwise tries to guess at
the age-group labels.

It is common for age groups to be coded using only the first
year, so that, for instance, \code{0}, \code{1}, \code{5},
\code{...}, \code{100} is short for \code{"0"},
\code{"1-4"}, \code{"5-9"}, \code{...}, \code{"100+"}.
\code{cleanAgeGroup} interprets a vector\code{age}
as a set of age-group labels if the unique values from \code{age},
once sorted, form a sequence, \code{0, 5, 10, ...} or
\code{0, 1, 5, 10, ...}.

\code{cleanAgeGroup} does the reformatting. \code{cleanAgeGroupConc}
constructs a \code{{Concordance}} between the old and
new age-group labels.

\code{cleanAgeGroup} is suitable for interactive use, or for
one-off analyses, where the data will not change in future.

Using \code{cleanAgeGroupConc} may be safer in production
code where the data may change. The workflow is as follows:
\enumerate{
 \item Use a vector of age labels to construct a concordance.
 \item Store the concordance.
 \item When new data arrives, use function
\code{\link{translate}}, plus the stored concordance,
  to clean the age data.
}
The advantange of this work flow is that \code{translate}
will throw an error if it strikes an age label that it
does not recognize.
}
\examples{
x1 <- c("0 - 4 years", "90 plus years", "25 - 29 years")
x2 <- c("10 to 19 Yr", "80 years and over")
x3 <- c(10, 0, 5, 1, 5)
x4 <- c("5--9", "10plus", "0--4")
x5 <- c("10plus", "10plus", "5--9", "5--9")

## Use cleanAgeGroup directly
cleanAgeGroup(x1)
cleanAgeGroup(x2)
cleanAgeGroup(x3)

## Set up concordance and use that
conc <- cleanAgeGroupConc(x4)
translate(x5, concordance = conc)

}
\seealso{
\code{\link{ageToAgeGroup}} converts exact
ages into age groups.
}
