% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Concordance-methods.R
\docType{methods}
\name{codes}
\alias{codes}
\alias{codes,Concordance-method}
\title{Extract the codes used by a concordance.}
\usage{
codes(object, classification)

\S4method{codes}{Concordance}(object, classification)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{Concordance}}.}

\item{classification}{Name of a classification.}
}
\value{
A character vector.
}
\description{
Extract the codes for one of the classifications used by an object of
class \code{\linkS4class{Concordance}}.
}
\examples{
x <- data.frame(code = 1:3,
                descriptor = c("North Island", "South Island",
                               "Stewart Island"))
x <- Concordance(x)
x
codes(x, classification = "code")
codes(x, classification = "descriptor")

x <- data.frame(sex1 = c("girl", "boy", "woman", "man"),
                sex2 = c("female", "male"))
x <- Concordance(x)
x
codes(x, classification = "sex1")
codes(x, classification = "sex2")
}
\seealso{
Names of classifications can be extracted or changed using
\code{\link{classifications}}.
}
