% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Concordance-methods.R
\name{codesAmbiguous}
\alias{codesAmbiguous}
\alias{codesAmbiguous,Concordance-method}
\title{Test whether codes in a concordance have fixed meanings.}
\usage{
codesAmbiguous(object)

\S4method{codesAmbiguous}{Concordance}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{Concordance}}.}
}
\value{
Logical.
}
\description{
Test whether any codes in an object of class
\code{\linkS4class{Concordance}} are ambiguous.  A code is ambiguous if it
appears in more than one classification, unless it only ever maps on to
itself.  See below for examples.
}
\examples{
x <- cbind(c1 = c("a", "b", "c"), c2 = c("x", "y", "x"))
x <- Concordance(x)
## no codes found in both classifications
codesAmbiguous(x)

x <- cbind(c1 = c("a", "b", "c"), c2 = c("a", "y", "x"))
x <- Concordance(x)
## "a" found in both classifications, but maps on to itself.
codesAmbiguous(x)

x <- cbind(c1 = c("a", "b", "c"), c2 = c("x", "a", "x"))
x <- Concordance(x)
## "a" found in both classifications, mapping on to different values.
codesAmbiguous(x)
}
\seealso{
The codes used by a concordances can be extracted using
\code{\link{codes}}.
}
