% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Concordance-methods.R
\name{coercion}
\alias{coercion}
\alias{as.matrix,Concordance-method}
\alias{as.data.frame,Concordance-method}
\alias{as.list,ManyToOne-method}
\title{Coerce concordances to other formats.}
\usage{
\S4method{as.matrix}{Concordance}(x, ...)

\S4method{as.data.frame}{Concordance}(
  x,
  row.names = NULL,
  optional = FALSE,
  stringsAsFactors = default.stringsAsFactors(),
  ...
)

\S4method{as.list}{ManyToOne}(x, ...)
}
\arguments{
\item{x}{A \code{\linkS4class{Concordance}}}

\item{\dots}{additional arguments to be passed to or from methods.}

\item{row.names}{\code{NULL} or a character vector giving the row names for
the data frame.}

\item{optional}{Logical.  See \code{\link[base]{as.data.frame}}.}

\item{stringsAsFactors}{Whether codes should be represented as factors.}
}
\value{
An matrix, data.frame, or list.
}
\description{
Coerce concordances to other formats.
}
