% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\docType{methods}
\name{collapseCategories}
\alias{collapseCategories}
\alias{collapseCategories,Counts,ANY,ANY,ANY,missing-method}
\alias{collapseCategories,Counts,ANY,missing,missing,OneToOne-method}
\alias{collapseCategories,Counts,ANY,missing,missing,ManyToOne-method}
\alias{collapseCategories,Values,ANY,ANY,ANY,missing-method}
\alias{collapseCategories,Values,ANY,missing,missing,ManyToOne-method}
\alias{collapseCategories,Values,ANY,missing,missing,OneToOne-method}
\title{Collapse multiple categories into a single category.}
\usage{
collapseCategories(object, dimension = NULL, old, new, concordance,
  weights = NULL, ...)

\S4method{collapseCategories}{Counts,ANY,ANY,ANY,missing}(object,
  dimension = NULL, old, new, weights = NULL)

\S4method{collapseCategories}{Counts,ANY,missing,missing,OneToOne}(object,
  dimension = NULL, concordance, weights = NULL)

\S4method{collapseCategories}{Counts,ANY,missing,missing,ManyToOne}(object,
  dimension = NULL, concordance, weights = NULL)

\S4method{collapseCategories}{Values,ANY,ANY,ANY,missing}(object, dimension,
  old, new, weights)

\S4method{collapseCategories}{Values,ANY,missing,missing,ManyToOne}(object,
  dimension, concordance, weights)

\S4method{collapseCategories}{Values,ANY,missing,missing,OneToOne}(object,
  dimension = NULL, concordance, weights = NULL)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Character or numeric vector giving the name or index of the
dimension where the categories are found - or names or indices if more than
one dimension is being changed.}

\item{old}{Character vector containing the names of the categories to be
collapsed.}

\item{new}{Character vector of length 1 giving the name of the new category
formed from the old categories.}

\item{concordance}{Object of class \code{"\linkS4class{Concordance}"}.}

\item{weights}{Object of class \code{\linkS4class{Counts}}.}

\item{\dots}{Not currently used.}
}
\description{
Merge multiple categories within a dimension or dimensions of
a \code{\linkS4class{DemographicArray}} object.
}
\details{
The categories to be collapsed, and the categories that result from collapsing can
be described through \code{old} and \code{new}, or through a \code{"\linkS4class{ManyToOne}"}
concordance:  see below for examples.

Pairs of of dimensions with dimtypes \code{origin} and \code{destination} or \code{parent}
and \code{child} can be selected by their base name: see the example below.
If no dimension is supplied, \code{collapseCategories} attempts to modify all dimensions.

\code{weights} are required if \code{object} has class
\code{\linkS4class{Values}} and are prohibited if
\code{object} has class \code{\linkS4class{Counts}}.
}
\examples{
titanic <- Counts(Titanic, dimtypes = c(Age = "state"))
collapseCategories(titanic, dimension = "Class",
                   old = c("1st", "2nd", "3rd"),
                   new = "Passengers")
## can also use the dimension index, though the name is clearer
collapseCategories(titanic, dimension = 1,
                   old = c("1st", "2nd", "3rd"),
                   new = "Passengers")
## set up and use a concordance
library(classconc)
conc <- data.frame(status = c("1st", "2nd", "3rd", "Crew"),
                   wealth = c("Rich", "Mixed", "Poor", "Mixed"))
conc <- Concordance(conc)
collapseCategories(titanic, dimension = "Class", concordance = conc)

library(demdata)
rates <- Values(VADeaths2)
popn <- Counts(VAPopn)
collapseCategories(rates, dimension = "sex", old = c("Male", "Female"),
                   new = "Combined", weights = popn)

occ <- occupationalStatus
names(dimnames(occ)) <- c("occupation_parent", "occupation_child")
occ <- Counts(occ)
## select a pair of dimensions by specifying a base name
collapseCategories(occ, dimension = "occupation",
                   old = c("2", "5"), new = "9")
## no dimension specified, so changes made to all dimensions
collapseCategories(occ, old = c("2", "5"), new = "9")
}
\seealso{
\code{\link{collapseIntervals}}, \code{\link{collapseIterations}},
\code{\link{collapseOrigDest}}, \code{\link{collapseDimension}},
\code{\link{dimtypes}}, \code{\link{dimscales}}
}
