% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Component-methods.R,
%   R/Counts-methods.R, R/Values-methods.R
\docType{methods}
\name{collapseDimension}
\alias{collapseDimension}
\alias{collapseDimension,Pool,ANY,ANY,missing-method}
\alias{collapseDimension,Counts,ANY,ANY,missing-method}
\alias{collapseDimension,Counts,ANY,ANY,Counts-method}
\alias{collapseDimension,Values,ANY,ANY,ANY-method}
\alias{collapseDimension,Values,ANY,ANY,missing-method}
\title{Collapse one or more dimensions of a demographic array.}
\usage{
collapseDimension(object, dimension = NULL, margin = NULL, weights, ...)

\S4method{collapseDimension}{Pool,ANY,ANY,missing}(object, dimension = NULL,
  margin = NULL, weights)

\S4method{collapseDimension}{Counts,ANY,ANY,missing}(object, dimension = NULL,
  margin = NULL, weights)

\S4method{collapseDimension}{Counts,ANY,ANY,Counts}(object, dimension = NULL,
  margin = NULL, weights)

\S4method{collapseDimension}{Values,ANY,ANY,ANY}(object, dimension = NULL,
  margin = NULL, weights)

\S4method{collapseDimension}{Values,ANY,ANY,missing}(object, dimension = NULL,
  margin = NULL)
}
\arguments{
\item{object}{An object of class' \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Names or indices of the dimensions to be collapsed.}

\item{margin}{Names or indices of the dimensions that will remain after the
collapsing.}

\item{weights}{Object of class \code{\linkS4class{Counts}}.
\code{weights} can only be used if \code{object} has class
\code{\linkS4class{Values}}.  It must have at least as much detail as
\code{object}.}

\item{\dots}{Not currently used.}
}
\value{
An object of the same class as \code{object}, but with fewer
dimensions, or a single number.
}
\description{
Aggregate values along one or more dimensions of an object of class
\code{\linkS4class{DemographicArray}}.  The aggregated
dimensions are dropped.
}
\details{
Either \code{dimension} or \code{margin} must be supplied, but not both.

\code{weights} are required if \code{object} has class
\code{\linkS4class{Values}}, and prohibited if \code{object} has class
\code{\linkS4class{Counts}}.

If all dimensions are collapsed, the result is equivalent to
\code{sum(object)}.

As discussed in \code{\link{dimtypes}}, some dimensions come in pairs.
Members of a pair have the same base name but different suffixes, e.g.
\code{"region_orig"} and \code{"region_dest"}.  Including the base name,
e.g. \code{"region"}, in the \code{dimension} or \code{margin} arguments has
the same effect as including both names.  If only one member of a pair of
dimensions is collapsed, the suffix is removed and from the name of the
remaining dimension, and the dimtype is changed to \code{"state"}.  See
below for examples.
}
\examples{
library(demdata)
## collapse Counts object
popn <- Counts(VAPopn)
popn
collapseDimension(popn, dimension = "age")
collapseDimension(popn, dimension = c("sex", "color"))
collapseDimension(popn, margin = c("sex", "color"))

## collapse Values object, supplying weights
rate <- Values(VADeaths2)
collapseDimension(rate, dimension = "age", weights = popn)

## special treatment of origin-destination dimensions
mig <- Counts(nz.mig)
mig
collapseDimension(mig, dimension = "island")
collapseDimension(mig, margin = "island")
## note dimension name and dimtype:
collapseDimension(mig, margin = "island_orig")

## collapse all dimensions
popn <- Counts(VAPopn)
collapseDimension(popn, dimension = c("age", "sex", "residence"))
collapseDimension(popn, margin = character())
}
\seealso{
\code{\link{addDimension}}, \code{\link{collapseCategories}}, %%
\code{\link{collapseIntervals}}
}
