% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{collapseIterations}
\alias{collapseIterations}
\alias{collapseIterations,DemographicArray-method}
\title{Summarize iterations.}
\usage{
collapseIterations(object, FUN = quantile, ...)

\S4method{collapseIterations}{DemographicArray}(object, FUN = quantile, ...)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{FUN}{A function.  Defaults to \code{\link{quantile}}, though others
can be used.}

\item{\dots}{Further arguments, passed to \code{FUN}.}
}
\description{
Collapse a dimension with \code{\link{dimtype}} \code{"iteration"},
typically replacing it with a much shorter dimension with dimtype
\code{"quantile"}.
}
\details{
Many estimation and forecasting methods produce sets of simulation runs or
iterations.  When graphing or reporting results, it is typically necessary
to reduce the full set of iterations to a set of summary values.  Function
\code{collapseIterations} carries out this reduction.  By default,
\code{collapseIterations} calculates quantiles.  However, users can supply
alternatives, including their own customized functions.

If \code{FUN} returns a vector, then the dimension with dimtype
\code{"iteration"} is replaced by one with dimtype \code{"quantile"} or
\code{"state"}.  If \code{FUN} returns a single value, then the
\code{"iteration"} dimension is dropped.

By default, function \code{\link{quantile}} uses quantiles \code{0, 0.25,
0.5, 0.75, 1}.  Quantiles more useful for summarizing iterations, such as
\code{0.025, 0.5, 0.975}, can be obtained by using a \code{prob} argument:
see below for an example.
}
\examples{
x <- Counts(array(rpois(n = 20, lambda = c(8, 12)),
                  dim = c(2, 10),
                  dimnames = list(sex = c("Female", "Male"),
                      iteration = 1:10)))
x
collapseIterations(x)

## pass different values for 'prob' to 'quantile'
collapseIterations(x, prob = c(0.025, 0.5, 0.975))

## summarize with a single value
collapseIterations(x, FUN = mean)

## a user-defined function
sumvals <- function(x) c(mean = mean(x), sd = sd(x))
collapseIterations(x, FUN = sumvals)
}
\seealso{
\code{\link{collapseCategories}}, \code{\link{collapseIntervals}},
\code{\link{collapseOrigDest}}, \code{\link{collapseDimension}},
\code{\link{dimtypes}}, \code{\link{dimscales}}
}

