% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicAccount-methods.R
\name{componentNames}
\alias{componentNames}
\alias{componentNames<-}
\alias{setComponentNames}
\alias{componentNames,DemographicAccount-method}
\alias{componentNames<-,DemographicAccount-method}
\alias{setComponentNames,DemographicAccount-method}
\title{Get or set the names of components of a demographic account.}
\usage{
componentNames(object)

componentNames(object) <- value

setComponentNames(object, value)

\S4method{componentNames}{DemographicAccount}(object)

\S4method{componentNames}{DemographicAccount}(object) <- value

\S4method{setComponentNames}{DemographicAccount}(object, value)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicAccount}}.}

\item{value}{Names of the components.}
}
\value{
\code{componentNames} returns a vector of names.
\code{setComponentNames} and the replacement form of
\code{componentNames} return a modified version of \code{object}.
}
\description{
A \code{\link[=DemographicAccount-class]{demographic account}} contains
counts of population, plus one or more components such as births,
deaths, or internal migration.  \code{componentNames} extracts the
names of the components.  \code{setNamesComponents} and
the replacement form of \code{namesComponents} can be used
to specify the names.
}
\examples{
population <- Counts(array(c(10, 15, 13, 16),
                           dim = c(2, 2),
                           dimnames = list(age = c("0-29", "30+"),
                                           time = c(1970, 2000))))
births <- Counts(array(13,
                       dim = c(1, 1),
                       dimnames = list(age = "30+",
                                       time = "1971-2000")))
deaths <- Counts(array(c(0, 9),
                       dim = c(2, 1),
                       dimnames = list(age = c("0-29", "30+"),
                                       time = c("1971-2000"))))
account <- Movements(population = population,
                     births = births,
                     exits = list(deaths = deaths))

componentNames(account)
componentNames(account) <- c("Births", "Deaths") # capitalized
componentNames(account)
library(dplyr) # to get pipes
account <- account \%>\%
setComponentNames(value = c("BIRTHS", "DEATHS"))
account
}
\seealso{
The names can also be set when creating
the account using functions \code{\link{Movements}} and
\code{Transitions}. Components can be extracted using
function \code{\link{components}}.
}
