% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicAccount-methods.R
\docType{methods}
\name{components}
\alias{components}
\alias{components,DemographicAccount-method}
\title{Extract one or more components from a demographic account.}
\usage{
components(object, names = NULL, simplify = FALSE)

\S4method{components}{DemographicAccount}(object, names = NULL,
  simplify = TRUE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicAccount}}.}

\item{names}{Names of the components.  If omitted all components
are returned.}

\item{simplify}{Logical. If \code{simplify} is \code{TRUE} (the default),
and if only one component is requested, then the return value for
\code{components} is the Counnts object itself, rather than a
named list of length 1.}
}
\value{
A single \code{\linkS4class{Counts}} object, or a named
list of Counts objects.
}
\description{
A \code{\link[=DemographicAccount-class]{demographic account}} contains
counts of population, plus one or more components such as births,
deaths, or internal migration.  \code{components} can be used to
extract one or more of these components, as a named list.
}
\examples{
population <- Counts(array(c(10, 15, 13, 16),
                           dim = c(2, 2),
                           dimnames = list(age = c("0-29", "30+"),
                                           time = c(1970, 2000))))
births <- Counts(array(13,
                       dim = c(1, 1),
                       dimnames = list(age = "30+",
                                       time = "1971-2000")))
deaths <- Counts(array(c(0, 9),
                       dim = c(2, 1),
                       dimnames = list(age = c("0-29", "30+"),
                                       time = c("1971-2000"))))
account <- Movements(population = population,
                     births = births,
                     exits = list(deaths = deaths))

components(account)
components(account, names = "births") ## Counts object
components(account, names = "births", simplify = FALSE) ## named list
components(account, names = c("deaths", "births"))
}
\seealso{
Population counts can be extracted using function
\code{\link{population}}.
}
