% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/Concordance-methods.R
\docType{class}
\name{Concordance-class}
\alias{Concordance-class}
\alias{ManyToOne-class}
\alias{OneToOne-class}
\alias{show,Concordance-method}
\title{S4 class for holding concordances.}
\usage{
\S4method{show}{Concordance}(object)
}
\arguments{
\item{object}{An object of class \code{Concordance}}
}
\description{
An object of class \code{Concordance} describes how codes from one
classification map on to codes from another classification.
The superclass \code{Concordance} has two subclasses:
\code{ManyToOne} and \code{OneToOne}.
}
\section{Slots}{

\describe{
\item{\code{values}}{A character matrix.}

\item{\code{classifications}}{A character vector of length 2.}
}}

\seealso{
Objects are created using function \code{\link{Concordance}}.
}
