% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance-generator.R
\name{Concordance}
\alias{Concordance}
\title{Create a concordance.}
\usage{
Concordance(object)
}
\arguments{
\item{object}{A matrix or data frame with two columns and unique colnames.}
}
\value{
An object of class \code{"\linkS4class{Concordance}"}.
}
\description{
Create an object of class \code{"\linkS4class{Concordance}"}.
}
\details{
\code{object} should contain two classifications, one in each column.  The
names of the classifications are taken from \code{colnames(object)}.
Duplicate rows are deleted.  The classifications must have a many-to-one or
one-to-one relationship.
}
\examples{
x <- cbind(c1 = c("a", "b", "c"), c2 = c("x", "y", "x"))
x <- Concordance(x)
classifications(x)

x <- cbind(c1 = c("a", "b", "c"), c2 = c("x", "y", "x"))
x <- Concordance(x)
x
classifications(x)
}
\seealso{
\code{"\linkS4class{Concordance}"}, \code{classifications}
}
