% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counts-generator.R, R/Values-generator.R
\name{countsone-valuesone}
\alias{CountsOne}
\alias{ValuesOne}
\alias{countsone-valuesone}
\title{Create one-dimensional object of class "Counts" or "Values"}
\usage{
CountsOne(values, labels, name, dimtype = NULL, dimscale = NULL)

ValuesOne(values, labels, name, dimtype = NULL, dimscale = NULL)
}
\arguments{
\item{values}{Numeric vector.}

\item{labels}{Character vector of labels.}

\item{name}{Name of dimension.}

\item{dimtype}{\code{\link{dimtype}} of dimension.  Optional.}

\item{dimscale}{\code{\link{dimscale}} of dimension.  Optional.}
}
\value{
Object of class \code{"\linkS4class{Counts}"} or
\code{"\linkS4class{Counts}"}.
}
\description{
Convenience functions for creating one-dimensions objects of class
\code{"\linkS4class{Counts}"} or \code{"\linkS4class{Values}"}.
}
\details{
When creating one-dimensional \code{"\linkS4class{Counts}"} or
\code{"\linkS4class{Counts}"} objects, using \code{CountsOne} and
\code{ValuesOne} is slightly easier than using the standard generator
functions \code{\link{Counts}} and \code{\link{Values}}.

If \code{values} is shorter than \code{labels}, it is recycled.
}
\examples{

CountsOne(values = c(-1, 1), labels = c("Female", "Male"), name = "sex")

ValuesOne(values = rnorm(10), labels = 0:9, name = "age")

ValuesOne(values = rnorm(3), labels = 1:3, name = "time", dimtype = "state")

ValuesOne(values = rnorm(10), labels = 0:9, name = "age", dimscale = "Points")
}
\seealso{
\code{\link{Counts}} and \code{\link{Values}} are used to created
\code{"\linkS4class{Counts}"} or \code{"\linkS4class{Counts}"} of any
dimension.
}

