% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CountsWithSubtotals-class}
\alias{CountsWithSubtotals-class}
\title{S4 Class For Counts with Subtotals.}
\description{
A \code{\linkS4class{Counts}} object with a set of subtotals attached.
Functions such as \code{\link{impute}} respect these subtotals.
}
\details{
The counts must have missing values, and the subtotals must not.
All non-missing values must be non-negative integers.
}
\section{Objects from the Class}{

Objects are typically created by calls to \code{\link{attachSubtotals}}.
}

\seealso{
\code{impute}
}
