% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{credibleInterval}
\alias{credibleInterval}
\alias{credibleInterval,DemographicArray-method}
\title{Calculate credible intervals}
\usage{
credibleInterval(object, width = 95, na.rm = FALSE)

\S4method{credibleInterval}{DemographicArray}(object, width = 95,
  na.rm = FALSE)
}
\arguments{
\item{object}{A \code{\link{DemographicArray}}. The array
must have a dimension with \code{\link{dimtype}}
\code{"iteration"}.}

\item{width}{A percentage value.}

\item{na.rm}{\code{TRUE} (the default) or \code{FALSE}.}
}
\value{
An object with the same class as \code{object},
in which the \code{"iteration"} dimension is replaced
by a \code{"quantiles"} dimension of length 2.
}
\description{
Calculate credible intervals from a \code{\link{DemographicArray}}.
The demographic array must have a dimension with \code{\link{dimtype}}
\code{"iteration"}.
}
\details{
\code{width} must be a percentage value. It can equal 100,
but cannot equal 0.

The credible intervals are calculated using the quantiles
\code{(q/2, 1 - q/2)} where \code{q = 1 - width/100}.

When \code{na.rm} is \code{TRUE} (the default), missing values
are removed before the credible interval is calculated.
}
\examples{
x <- Values(array(runif(20),
                  dim = c(2, 10),
                  dimnames = list(sex = c("F", "M"),
                                  iteration = 1:10)))
credibleInterval(x)
credibleInterval(x, width = 50)
x[1] <- NA
credibleInterval(x)
credibleInterval(x, na.rm = FALSE)
}
\seealso{
\code{credibleInterval} calls function
\code{\link{collapseIteration}}.
}
