% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DemographicAccount-class}
\alias{DemographicAccount-class}
\alias{Movements-class}
\alias{Transitions-class}
\title{Classes "DemographicAccount", "Movements", "Transitions"}
\description{
Classes for describing demographic accounts.
}
\details{
There are two subclasses: Movements and Transitions
}
