% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DimScale-methods.R,
%   R/MetaData-methods.R
\name{dimscales}
\alias{dimscales}
\alias{dimscale}
\alias{dimscales<-}
\alias{dimscales,DimScale-method}
\alias{dimscales,MetaData-method}
\alias{dimscales<-,MetaData-method}
\title{Get or set the dimscales of a demographic array}
\usage{
dimscales(object, use.names = TRUE)

dimscales(object) <- value

\S4method{dimscales}{DimScale}(object)

\S4method{dimscales}{MetaData}(object, use.names = TRUE)

\S4method{dimscales}{MetaData}(object) <- value
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{use.names}{Logical.  Whether to include the names of the dimensions in
the return value.}

\item{value}{Character vector of valid dimscales.  See table below for a
list of valid dimscales.}
}
\value{
\code{dimscales} returns a character vector with the dimscales of
each dimension.

The replacement method changes the dimscales of \code{object}.
}
\description{
Query or (to a limited extent) change the dimscales of a demographic array.
}
\details{
Each dimension of a \code{\linkS4class{DemographicArray}} object has a
'dimscale' describing the measurement scale used by that dimension.
Examples include \code{"Categories"}, \code{"Intervals"}, and
\code{"Points"}. See below for details.

Every dimension of a \code{\linkS4class{DemographicArray}} object has a
dimscale that identifies the measurement scale used by that dimension.

The table below lists valid dimscales.  Dimscales \code{"Categories"},
\code{"Intervals"}, and \code{"Points"} are used much more often than the
other three, which are associated with particular \code{\link{dimtypes}}.

\tabular{lll}{ \emph{dimscale} \tab \emph{Description} \tab \emph{Example}
\cr
\code{"Categories"} \tab Discrete qualitative categories.  No
duplicates. \tab \code{"East", "West"} \cr
\code{"Sexes"} \tab Biological sexes.  \tab \code{"Female", "Male"} \cr
\code{"Intervals"} \tab' Segments of real line, ordered, with no gaps
or overlap.  First interval can
extend to negative infinity, and last interval can extend to positive
infinity. \tab \code{"0", "1-4", "5-9", "10-14", "15+"}. \cr
\code{"Points"} \tab Real numbers, ordered. \tab \code{2000, 2005, 2010} \cr
\code{"Iterations"} \tab Positive integers, ordered.  Used exclusively with
dimtype \code{"iteration"}. \tab \code{1, 2, 3} \cr
\code{"Quantiles"} \tab
Real numbers in interval [0,1], ordered.  Used exclusively with dimtype
\code{"quantile"}. \tab \code{"2.5\%", "50\%", "97.5\%"} \cr 
\code{"Triangles"} \tab
Upper or lower Lexis triangle.  Used exclusively with dimtype
\code{"triangle"}. \tab \code{"Lower", "Upper"} \cr 
\code{"Pool"} \tab
In-migration or out-migration in a 'migrant pool' model.  Used exclusively
with dimtype \code{"pool"}. \tab \code{"Ins", "Outs"} \cr }
}
\note{
Changing the dimscales of an existing object can be tricky, since the
new dimscales must be consistent with the existing dimtypes.  If major
changes to the dimtypes and dimscales of an object are required, it may be
easiest to create the object again from scratch.
}
\examples{
library(demdata)
x <- Counts(VADeaths2)
dimscales(x)

a <- array(rnorm(10),
           dim = c(5, 2),
           dimnames = list(year = 2000:2004, sex = c("Female", "Male")))
x <- Values(a, dimscales = c(year = "Intervals"))
dimscales(x)
dimscales(x)[1] <- "Points"
dimscales(x)
}
\seealso{
The type of information encoded by a dimension is represented by
\code{\link{dimtypes}}.

To change elements within a dimension, use functions such as
\code{\link{collapseCategories}}. To add extra elements, use
\code{\link{dbind}}.
}
