% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/MetaData-methods.R
\name{dimtypes}
\alias{dimtypes}
\alias{dimtype}
\alias{dimtypes<-}
\alias{dimtypes,MetaData-method}
\alias{dimtypes<-,MetaData-method}
\title{Get or set the dimtypes of an object.}
\usage{
dimtypes(object, use.names = TRUE)

dimtypes(object) <- value

\S4method{dimtypes}{MetaData}(object, use.names = TRUE)

\S4method{dimtypes}{MetaData}(object) <- value
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{use.names}{Logical.  Whether to include the names of the dimensions in
the return value.}

\item{value}{Character vector of valid dimtypes.  See table below for a list
of valid dimtypes.}
}
\value{
\code{dimtypes} returns a character vector with the dimtypes of each
dimension.

The replacement method changes the dimtypes of \code{object}.
}
\description{
Query or (to a limited extent) change the dimtypes of an object.
}
\details{
Each dimension of a \code{\linkS4class{DemographicArray}} object has a
'dimtype' describing the type of information held by that dimension.
Examples include \code{"age"}, \code{"time"}, and \code{"state"}. See below
for details.

Every dimension of a \code{\linkS4class{DemographicArray}} object has a
dimtype.  A dimtype identifies the type of variable encoded by the
dimension.

The table below lists valid dimtypes.  The dimtypes from \code{"age"} to
\code{"destination"} are the most commonly-encountered, with the remaining
dimtypes being more specialized.  Different dimtypes require different
\code{\link{dimscales}}.

\tabular{rll}{ dimtype \tab Description \tab Permitted dimscales \cr
\code{"state"} \tab Any qualitative attribute \tab \code{"Categories"} \cr
\code{"age"} \tab Age \tab \code{"Intervals"}, \code{"Points"} \cr
\code{"sex"} \tab Sex \tab \code{"Sexes"} \cr
\code{"time"} \tab Periods or exact times \tab \code{"Intervals"},
\code{"Points"} \cr
\code{"origin"}, \code{"destination"} \tab Starting and finishing states
\tab \code{"Categories"} \cr
\code{"parent"}, \code{"child"} \tab Parent's state versus child's state
\tab \code{"Categories"} \cr
\code{"cohort"} \tab Cohort \tab \code{"Intervals"} \cr
\code{"iteration"} \tab Simulation number \tab \code{"Iterations"} \cr
\code{"quantile"} \tab Quantile (between 0 and 1) \tab \code{"Quantiles"}
\cr \code{"triangle"}
\tab Lexis triangle \tab \code{"Triangle"} \cr }

\itemize{
  \item \code{"state"}.  The default dimtype, used for any attribute
that is not dealt with by the more specialized dimtypes.
\item \code{"age"}.  Age, measured on a numerical scale Can be 'exact' ages, or
age-groups. If age is measured qualitatively (eg \code{"children"} and
\code{"adults"}), then dimtype \code{"state"} should be used instead. An
object can have at most one dimension with dimtype \code{"age"}.
  \item \code{"sex"}.  Biological sex.
  \item \code{"time"}.  Time, measured on a numerical scale.  Can be points in time,
or intervals between points in time. If time is measured qualitatively (eg
\code{"before"} and \code{"after"}), then dimtype \code{"state"} should be
used instead.
  \item \code{"origin"}, \code{"destination"}.  Starting and
finishing values for an attribute that is subject to change, such as country
of residence.  \code{"origin"} and \code{"destination"} dimensions must come
in pairs with a common base name and suffixes \code{"_orig"} and
\code{"_dest"}, each \code{"country_orig"} and \code{"country_dest"}.  There
is no limit to the number of origin-destination pairs that an object may
have.
  \item \code{"parent"}, \code{"child"}.  Parents' attributes versus
those of their children.  Used to model transmission of attributes such as
ethnicity from parents to children. \code{"parent"} and \code{"child"}
dimensions must come in pairs with a common base name and suffixes
\code{"_parent"} and \code{"_child"}, e.g. \code{"ethnicity_parent"} and
\code{"ethnicity_child"}.  There is no limit to the number of parent-child
pairs that an object may have.
  \item \code{"iteration"}.  Iteration number
from a simulation.  An object can have at most one dimension with dimtype
\code{"iteration"}.  If an object has a dimension with dimtype
\code{"iteration"} it may not have a dimension with dimtype
\code{"quantile"}.
  \item \code{"quantile"}.  Sample quantiles, typically
summarizing simulation results.  An object can have at most one dimension
with dimtype \code{"quantile"}.  If an object has a dimension with dimtype
\code{"quantile"} it may not have a dimension with dimtype
\code{"iteration"}.
  \item \code{"triangle"}.  Lexis triangle.
}
}
\note{
Changing the dimtype of a dimension in an existing object can be
tricky, since the dimension may have \code{\link{dimscales}} that are not
permitted for the new dimtype.  If major changes to the dimtypes and
dimscales of an object are required, it may be easiest to create the object
again from scratch, perhaps using \code{\link{as.array}} on the object
first.
}
\examples{
library(demdata)
x <- Counts(VADeaths2)
dimtypes(x)

a <- array(1:8,
           dim = c(2, 2, 2),
           dimnames = list(age = c("0-39", "40+"),
                           tri = c("Upper", "Lower"),
                           period = c("1961-2000", "2001-2040")))
x <- Counts(a)
dimtypes(x)
dimtypes(x)[2] <- "triangle"
dimtypes(x)
}
\seealso{
\code{\link{dimscales}}
}
