% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame-methods.R
\name{dtabs}
\alias{dtabs}
\title{Simplified version of 'xtabs' for constructing
demographic arrays}
\usage{
dtabs(data, formula, na.rm = FALSE, fill = 0L)
}
\arguments{
\item{data}{A data.frame or matrix.}

\item{formula}{A formula: see \code{\link{xtabs}} for details.}

\item{na.rm}{Logical. Whether to remove \code{NA}s before tabulating.
Defaults to \code{FALSE}.}

\item{fill}{The value to use for combinations of
variables that do not occur in \code{data}.}
}
\value{
An array.
}
\description{
\code{dtabs} is a simplified version of function
\code{\link[stats]{xtabs}}, designed specifically for
constructing demographic arrays.
}
\details{
The \code{data} argument comes first, so that \code{dtabs}
works nicely with pipes.

The \code{fill} argument makes it easy to control what
value used for combinations of classifying variables
that do not occur in the original dataset \code{data}.
See below for examples.

The return value is a plain \code{\link{array}}, not an
\code{\link{xtabs}} object.

Unlike \code{xtabs}, \code{dtabs} has no \code{subset} argument.
Rather than being combined with the tabulation, subsetting
should be separated out into its own operation, via a
function such as \code{\link{subset}} or \code{\link[dplyr]{filter}}.
}
\examples{
d <- data.frame(age = c("young", "old", "young", "old"),
                sex = c("Female", "Female", "Male", "Male"),
                count = 1:4)
dtabs(d, count ~ age + sex)
dtabs(d, count ~ age)
dtabs(d, ~ age + sex)
dtabs(d, ~ age)

d_incomplete <- data.frame(age = c("young", "old", "young"),
                           sex = c("Female", "Female", "Male"),
                           count = 1:3)
## default value of fill is 0
dtabs(d_incomplete, count ~ age + sex)
dtabs(d_incomplete, count ~ age + sex, fill = NA)
}
\seealso{
\code{dtabs} is based on \code{\link{xtabs}}. To turn
a plain array created by \code{dtabs} into a demographic array,
use function \code{\link{Counts}} or \code{\link{Values}}.
}
