% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\name{expandCategories}
\alias{expandCategories}
\alias{expandCategories,Counts,ANY,missing,missing,ManyToOne-method}
\alias{expandCategories,Values,ANY,ANY,ANY,missing-method}
\alias{expandCategories,Values,ANY,missing,missing,ManyToOne-method}
\title{Expand a single category into multiple categories.}
\usage{
expandCategories(
  object,
  dimension = NULL,
  old,
  new,
  concordance,
  weights = NULL,
  means = FALSE,
  n = NULL
)

\S4method{expandCategories}{Counts,ANY,missing,missing,ManyToOne}(
  object,
  dimension,
  concordance,
  weights = NULL,
  means = FALSE,
  n = NULL
)

\S4method{expandCategories}{Values,ANY,ANY,ANY,missing}(object, dimension, old, new, weights)

\S4method{expandCategories}{Values,ANY,missing,missing,ManyToOne}(object, dimension, concordance, weights)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Character or numeric vector giving the name or index of the
dimension where the categories are found - or names or indices if more than
one dimension is being changed.}

\item{old}{Character vector of length 1 containing the name of the
category to be expanded.}

\item{new}{Character vector giving the names of the new categories.}

\item{concordance}{Object of class \code{"\linkS4class{ManyToOne}"}.}

\item{weights}{Object of class \code{\linkS4class{Counts}}.}

\item{means}{If \code{FALSE} (the default) the allocation is done randomly;
if \code{TRUE}, values are set equal to their means.}

\item{n}{Number of iterations to generate.  Ignored if \code{counts} or
\code{weights} has a dimension with \code{\link{dimtype}}
\code{"iteration"}.}
}
\description{
Using a \code{\linkS4class{ManyToOne}} concordance, expand the categories
used by one or more dimensions of a demographic array.
}
\seealso{
\code{\link{collapseCategories}}
}
