% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{expandIntervals}
\alias{expandIntervals}
\alias{expandIntervals,Counts,ANY,numeric,missing,missing,missing-method}
\alias{expandIntervals,Counts,ANY,missing,numeric,missing,missing-method}
\title{Split wide intervals into narrower ones}
\usage{
expandIntervals(object, dimension, breaks = NULL, width = NULL,
  old = NULL, weights, means = FALSE)

\S4method{expandIntervals}{Counts,ANY,numeric,missing,missing,missing}(object,
  dimension, breaks = NULL, width = NULL, old = NULL, weights,
  means = FALSE)

\S4method{expandIntervals}{Counts,ANY,missing,numeric,missing,missing}(object,
  dimension, breaks = NULL, width = NULL, old = NULL, weights)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Name or index of the dimension where the intervals are
found.}

\item{breaks}{Numeric vector giving the breaks between intervals after the
merging has occurred.}

\item{width}{The length of the intervals after the merging has occurred.}

\item{old}{The labels of the intervals to be merged.}

\item{weights}{Object of class \code{\linkS4class{Counts}} or the number
\code{1}.  Required when \code{object} has class
\code{\linkS4class{Values}} and prohibited when it has class
\code{\linkS4class{Counts}}.}

\item{means}{If \code{FALSE} (the default) the allocation is done randomly;
if \code{TRUE}, values are set equal to their means.}
}
\description{
Split a dimension with \code{\link{dimscale}} \code{"Intervals"} into smaller
units, and redistribute counts within these smaller units.
}
\details{
NOTE - THIS FUNCTION IS INCOMPLETE, AND WILL BE MODIFIED TO INCLUDE
ADDITIONAL ARGUMENTS.
}
