% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{exposure}
\alias{exposure}
\alias{exposure,Counts-method}
\title{Calculate exposure from population counts.}
\usage{
exposure(object, triangles = FALSE)

\S4method{exposure}{Counts}(object, triangles = FALSE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}}.}

\item{triangles}{Logical.  If \code{TRUE}, the return value
includes a dimension with \code{\link{dimtype}} code{"triangle"}.
Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\linkS4class{Counts}}.
}
\description{
\code{object} typically has a dimension with \code{\link{dimtype}}
\code{time} and \code{dimscale} \code{"Points"}, and a dimension
with dimtype \code{"age"} and dimscale \code{"Intervals"}.
The complete set of choices for age-time-cohort dimensions and
dimscales is
\tabular{ccc}{
  \code{time} \tab \code{age} \tab \code{cohort} \cr
  \code{Points} \tab \code{Intervals} \tab <none> \cr
  \code{Points} \tab <none> \tab <none> \cr
  <none> \tab \code{Points} \tab <none> \cr
  \code{Points} \tab <none> \tab \code{Intervals} \cr
  <none> \tab \code{Points} \tab \code{Intervals}
}
}
\details{
If \code{triangles} is \code{TRUE}, then exposure is calculated
separately for lower and upper Lexis triangles.  \code{triangles}
can only by \code{TRUE} if  \code{object} has dimensions with
\code{\link{dimtype}} \code{"time"} and \code{"age"}.
}

