% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{exposureBirths}
\alias{exposureBirths}
\alias{exposureBirths,Counts-method}
\title{Calculate exposure to use for modelling births.}
\usage{
exposureBirths(object, triangles = FALSE, births = NULL,
  dominant = c("Female", "Male"))

\S4method{exposureBirths}{Counts}(object, triangles = FALSE, births = NULL,
  dominant = c("Female", "Male"))
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}}.}

\item{triangles}{Logical.  If \code{TRUE}, the return value
includes a dimension with \code{\link{dimtype}} code{"triangle"}.
Defaults to \code{FALSE}.}

\item{births}{A \code{\linkS4class{DemographicArray}} object giving
birth counts or rates.}

\item{dominant}{\code{"Female"} (the default) or \code{"Male"}.
Whether to use the number of potential mothers or potential
fathers when calculating exposures.}
}
\value{
A \code{\linkS4class{Counts}} object, with the same
dimensions as \code{births}.
}
\description{
In models of births, the relevant exposure term is usually the person-years
lived by women of reproductive age.  \code{exposureBirths} makes it easy to
calculate this term in typical cases.
}
\details{
The \code{births} argument is a tabulation of births or birth rates, typically
including the age of mothers. Minimum and maxium ages for exposure are taken
from this tabulation.

Fertility models are almost always "female dominant", that is, they
only take account of the number of reproductive-age females, and ignore
the number of reproductive age-males.  By default, if \code{object}
has two sexes, then \code{exposureBirths} only includes females.
However, if \code{dominant} is set to "Male", it will only include
males.

\code{births} can contain a sex dimension.  However, this dimension
describes the sex of the child, not the parent.  The sex of the parent
is still determined by \code{dominant} argument, and the same
exposure is used for female and male births.
}
\examples{
## prepare data
library(demdata)
popn <- demdata::nz.popn.reg
births <- demdata::nz.births.reg
popn <- Counts(popn, dimscales = c(year = "Points"))
births <- Counts(births, dimscales = c(year = "Intervals"))
popn <- subarray(popn, year \%in\% c(1996, 2001, 2006, 2011, 2016))
births <- collapseDimension(births, dimension = "region")
births <- collapseIntervals(births, dimension = "year", width = 5)
births.nosex <- collapseDimension(births, dimension = "sex")

## no triangles
exposureBirths(popn, births = births.nosex)

## with triangles
exposureBirths(popn, triangles = TRUE, births = births.nosex)

## dominant
exposureBirths(popn, births = births.nosex, dominant = "Male")

## distinguish sex of births
exposureBirths(popn, births = births)
}
\seealso{
To calculate exposure for quantities other than births,
see \code{\link{exposure}} (and, for origin-destination or parent-child
arrays, possibly also \code{\link{addPair}}). To tidy fertility data,
see \code{\link{reallocateToEndAges}}.
}
