% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{fillForward}
\alias{fillForward}
\title{Use observed values to fill in missing values in vector}
\usage{
fillForward(x, firstBackward = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{firstBackward}{Logical. Whether to fill upwards
if values at start of \code{x} are missing.}
}
\value{
A filled-in version of \code{x}.
}
\description{
Fill in missing values in a vector by carrying forward
observed values.
}
\details{
If one or more values at the start of the vector
are missing, and if \code{firstBackward} is \code{TRUE},
then the missing values at the start are filled in
by carrying the first value backwards.

\code{fillForward} is based on Stack Overflow post
https://stackoverflow.com/questions/1782704
which in turn cites function \code{na.locf}
in package \code{zoo}.
}
\examples{
x <- c("A", NA, NA, "B", NA)
fillForward(x)

## first values missing
x <- c(NA, NA, 1, NA, NA, 2, NA)
fillForward(x) # firstBackward TRUE by default
fillForward(x, firstBackward = FALSE)
}
