% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/CountsWithSubtotals-methods.R, R/DemographicArray-methods.R
\docType{methods}
\name{impute}
\alias{impute}
\alias{impute,CountsWithSubtotals-method}
\alias{impute,DemographicArray-method}
\title{Quick and dirty imputation of missing values.}
\usage{
impute(object, mult = NULL, max = NULL)

\S4method{impute}{CountsWithSubtotals}(object, mult = NULL, max = NULL)

\S4method{impute}{DemographicArray}(object, mult = NULL, max = NULL)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{mult}{If \code{TRUE}, a multiplicative model is used to impute values;
if \code{FALSE}, an additive model is used.  Defaults to \code{TRUE} if all
non-missing values in \code{object} are non-negative.}

\item{max}{Numeric vector giving maximum values for \code{object}.  Imputed
values respect this maximum; existing values are unaffected.  \code{max} is
recycled so that it has the same length as \code{object}.}
}
\value{
An object with the same class as \code{object}, with the missing
values filled in.
}
\description{
Imputes missing values by fitting an additive or multiplicative main effects
model to the non-missing data, and then randomly drawing values for the
missing values, based on predictions from the model.
}
\details{
\code{impute} performs a single imputation based on a simple, and not
necessarily appropriate, model.  \code{impute} is useful for generating
starting values for iterative calculations, for tidying data with a small
proportion of missing values, or for rough approximations.  For more
sophisticated approaches to the imputation of demographic data see package
\pkg{demest}.

If \code{mult} is \code{TRUE}, meaning that the multiplicative model is
used, all non-missing values must be non-negative (0s are omitted from the
model.)  If the \code{mult} is \code{TRUE} all imputed values are guaranteed
to be non-negative.

If all non-missing values are non-negative integers, imputed values are
drawn from a Poisson distribution.  Other they are drawn from a normal
distribution.

The \code{max} argument is useful when \code{object} represents draws from a
binomial distribution.  Setting \code{max} to the number of trials for each
cell ensures that the imputed number of successes does not exceed the number
of trials.
}
\examples{
library(demdata)
popn <- Counts(VAPopn)
popn[sample(length(popn), size = 5)] <- NA
popn
impute(popn)
impute(popn, mult = TRUE)
impute(popn, max = 500)
}
\references{
Little, RJA and Rubin, DB. 2002. \emph{Statistical Analysis with
Missing Data}, Wiley.
}
\seealso{
\code{\linkS4class{DemographicArray}}, \code{\link{perturb}},
\code{\link{extrapolate}}
}

