% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{internal-account}
\alias{internal-account}
\alias{Population-class}
\alias{Component-class}
\alias{Births-class}
\alias{BirthsMovements-class}
\alias{BirthsMovementsHasParentChild-class}
\alias{BirthsMovementsNoParentChild-class}
\alias{BirthsTransitions-class}
\alias{BirthsTransitionsHasParentChild-class}
\alias{BirthsTransitionsNoParentChild-class}
\alias{Internal-class}
\alias{InternalMovements-class}
\alias{InternalMovementsNet-class}
\alias{InternalMovementsOrigDest-class}
\alias{InternalMovementsPool-class}
\alias{InternalTransitions-class}
\alias{Entries-class}
\alias{EntriesMovements-class}
\alias{EntriesTransitions-class}
\alias{Exits-class}
\alias{ExitsMovements-class}
\alias{ExitsTransitions-class}
\alias{NetMovements-class}
\alias{Accession-class}
\alias{Exposure-class}
\title{Classes used by DemographicAccount.}
\description{
These classes are used internally by DemographicAccount.  They should never
be used directly by end users.  The only reason they are exported is to
allow other developers to extend classes "Counts" and "Values".
}
\keyword{internal}
