% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{intervalContainsTruth}
\alias{intervalContainsTruth}
\alias{intervalContainsTruth,DemographicArray,DemographicArray-method}
\alias{intervalContainsTruth,DemographicArray,numeric-method}
\title{Test whether intervals contain true values}
\usage{
intervalContainsTruth(interval, truth)


  \S4method{intervalContainsTruth}{DemographicArray,DemographicArray}(interval,
  truth)

\S4method{intervalContainsTruth}{DemographicArray,numeric}(interval, truth)
}
\arguments{
\item{interval}{A \code{\linkS4class{DemographicArray}},
with a quantile dimension of length 2.}

\item{truth}{A \code{\linkS4class{DemographicArray}}, or a single
number.}
}
\value{
An \code{\linkS4class{DemographicArray}}
of \code{1}s and \code{0}s with the same dimensions
as \code{truth}, or, if \code{truth} is a number,
a single \code{1} or \code{0}.
}
\description{
Test whether intervals contain specified values.  These values
are referred to as 'truth' because \code{intervalContainsTruth}
is typically used as part of a simulation study,
where the analyst generates the true values, and
tests the ability of the model to recover these values.
}
\details{
\code{interval} must have the same dimensions
and dimscales as \code{truth}, and also a dimension
with length 2 and \code{\link{dimtype}}
\code{"quantile"} specifying the intervals.

An interval \code{(l, u)} is said to contain a value \code{v}
if \code{l <= t <= u}. The use of \code{<=} is immaterial
when \code{l} and \code{u} are real numbers but makes a difference
when \code{l} and \code{u} are integers.

\code{truth} can be a single number, in which case all dimensions
of \code{interval}, other than the quantile dimension, should
have length 1.

\code{valueInInterval} is stricter about the compatibility
of its arguments than most functions in \code{dembase}.
Although it reorders dimensions and categories
in \code{interval} and \code{truth}, it does not collapse
or expand dimensions, or drop any levels.

When \code{truth} is a \code{\linkS4class{DemographicArray}}
\code{intervalContainsTruth} returns a \code{\linkS4class{Values}}
object with \code{1}s and \code{0}s,
rather than an ordinary array of
\code{TRUE}s and \code{FALSE}s. The advantage of the
\code{\linkS4class{Values}} object is that it can
more easily be collapsed or otherwise manipulated:
see below for an example.
}
\examples{
interval <- Values(array(c(-0.1, 0.4, 0.2, 0.7),
                         dim = c(2, 2),
                         dimnames = list(sex = c("Female", "Male"),
                                         quantile = c("2.5\%", "97.5\%"))))
truth <- ValuesOne(c(0.3, 0.5),
                   labels = c("Female", "Male"),
                   name = "sex")
intervalContainsTruth(interval = interval,
                      truth = truth)

interval <- ValuesOne(c(5L, 8L),
                      labels = c("5\%", "95\%"),
                      name = "quantile")
truth <- 5L
intervalContainsTruth(interval = interval,
                      truth = truth)
intervalContainsTruth(interval = interval,
                      truth = truth)
}
\seealso{
Function \code{\link{credibleInterval}} creates
\code{interval} objects of the right form
for \code{intervalContainsTruth}. Function \code{\link{MSE}},
is another function that may be useful in simulation studies.
}
