% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{intervalWidth}
\alias{intervalWidth}
\alias{intervalWidth,DemographicArray-method}
\title{Calculate the widths of credible intervals}
\usage{
intervalWidth(interval)

\S4method{intervalWidth}{DemographicArray}(interval)
}
\arguments{
\item{A}{\code{\link{DemographicArray}}.}

\item{A}{\code{\link{DemographicArray}}
holding the widths.}
}
\description{
Calculate the widths of credible intervals
stored in a \code{\link{DemographicArray}}.
If \code{(l, u)}, is a credible interval,
then the width of that interval equals
\code{u - l}.
}
\details{
\code{interval} must have a dimension of length
2 with \code{\link{dimtype}} \code{"quantile"}
specifying the intervals.  \code{interval} is
typically calculated using function
\code{\link{credibleInterval}}.
}
\examples{
#' set.seed(0)
object <- Values(array(runif(n = 20),
                       dim = c(2, 10),
                       dimnames = list(sex = c("F", "M"),
                                       iteration = 1:10)))
interval <- credibleInterval(object)
interval
intervalWidth(interval)
object <- Values(array(rpois(n = 20, lambda = 10),
                       dim = c(2, 10),
                       dimnames = list(sex = c("F", "M"),
                                       iteration = 1:10)))
interval <- credibleInterval(object, width = 90)
interval
intervalWidth(interval)
}
