% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R,
%   R/MetaData-methods.R
\name{limits}
\alias{limits}
\alias{limits,DemographicArray-method}
\alias{limits,MetaData-method}
\title{First and last categories used by each dimension.}
\usage{
limits(object, components = FALSE)

\S4method{limits}{DemographicArray}(object, components = FALSE)

\S4method{limits}{MetaData}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}
or \code{\linkS4class{DemographicAccount}}.}

\item{components}{Logical. Only used if \code{object} is a
\code{\linkS4class{DemographicAccount}}.  If \code{TRUE} then
limits are shown for all demographic series; if \code{FALSE}
(the default) limits are shown only for population.}
}
\value{
If \code{object} is a \code{\linkS4class{DemographicArray}},
or if \code{object} is a \code{\linkS4class{DemographicAccount}} and
\code{components} is \code{FALSE}, then the return value is a data.frame;
otherwise it is a list of data.frames.
}
\description{
Gives the first and last categories used by each dimension
of a \code{\linkS4class{DemographicArray}} object or a
\code{\linkS4class{DemographicAccount}}.  It provides a
quick way of understanding the structure of an object.
}
\examples{
library(demdata)
popn <- Counts(VAPopn)
limits(popn)
}
