% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R, R/MetaData-methods.R
\docType{methods}
\name{midpoints}
\alias{midpoints}
\alias{midpoints,DemographicArray,ANY-method}
\alias{midpoints,DemographicArray,missing-method}
\alias{midpoints,MetaData,integer-method}
\title{Replace intervals by their midpoints.}
\usage{
midpoints(object, dimension)

\S4method{midpoints}{DemographicArray,ANY}(object, dimension)

\S4method{midpoints}{DemographicArray,missing}(object)

\S4method{midpoints}{MetaData,integer}(object, dimension)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Names or indices of dimensions to be converted.  If
omitted, all dimensions with dimtype \code{"Intervals"} are converted.}
}
\value{
Object of class \code{\linkS4class{DemographicArray}}.
}
\description{
Change \code{"Intervals"} \code{\link{dimscales}} to \code{"Points"},
replacing each interval with its midpoint.  This can be useful for plotting
or calculating mean values, among other things.
}
\details{
\code{midpoints} is typically called by other functions, but it may
sometimes be useful to call it directly.
}
\examples{
library(demdata)
rates <- Values(VADeaths2)
midpoints(rates)

a <- array(1:4,
           dim = c(2, 2),
           dimnames = list(age = c("0-39", "40+"),
                           period = c("2001-2010", "2011-2020")))
x <- Counts(a)
midpoints(x)
midpoints(x, dimension = c("age", "period"))
midpoints(x, dimension = "period")
midpoints(x, dimension = 2)
}
\seealso{
\code{\link{dimscales}}, \code{\link{dplot}}
}

