% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Movements-generator.R
\name{Movements}
\alias{Movements}
\title{Create a demographic account based on movements.}
\usage{
Movements(population, births = NULL, internal = NULL, entries = list(),
  exits = list(), net = list())
}
\arguments{
\item{population}{A \code{\linkS4class{Counts}} object.}

\item{births}{A \code{\linkS4class{Counts}} object.}

\item{internal}{If \code{internal} uses an origin-destination format,
an ordinary \code{\linkS4class{Counts}} object; if it uses a pool or
net format, an object created by functions \code{\link{collapseOrigDest}},
\code{\link{Pool}}, or \code{\link{Net}}.}

\item{entries}{A named list of \code{\linkS4class{Counts}} objects.}

\item{exits}{A named list of \code{\linkS4class{Counts}} objects.}

\item{net}{A named list of \code{\linkS4class{Counts}} objects.}
}
\description{
The only compulsory argument is \code{population}.  All demographic series
must have a time dimension.  In the case of \code{population}, the time
dimension must have \code{\link{dimscale}} \code{"Points"}; in all other cases
it must have \code{\link{dimscale}} \code{"Intervals"}.
}

