% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\name{MSE}
\alias{MSE}
\alias{MSE,DemographicArray,DemographicArray-method}
\title{MSE (Mean Squared Error)}
\usage{
MSE(point, truth)

\S4method{MSE}{DemographicArray,DemographicArray}(point, truth)
}
\arguments{
\item{point}{A \code{\linkS4class{DemographicArray}} containing
the point estimates.}

\item{truth}{A \code{\linkS4class{DemographicArray}} containing
the true values.}
}
\value{
An object with the same class as \code{truth}.
}
\description{
Mean Squared Error, or MSE, measures the extent to which
point estimates (including forecasts) from a model come
close to the truth.  It is commonly used in simulation
studies, where the true values are generated by the
analyst, and hence known.
}
\details{
If \code{p} is a point estimate and \code{t} is the truth
then the MSE is \code{(p - t)^2}.

\code{MSE} is stricter about the compatibility
of its arguments than most functions in \code{dembase}.
Although it reorders dimensions and categories
in \code{point} and \code{truth}, it does not collapse
or expand dimensions, or drop any levels.
}
\examples{
point <- Values(array(c(10.2, 8.7, 3.4, 8.1),
                      dim = c(2, 2),
                      dimnames = list(region = c("A", "B"),
                                      sex = c("F", "M"))))
truth <- Values(array(c(9.8, 8.3, 3.7, 7.9),
                      dim = c(2, 2),
                      dimnames = list(region = c("A", "B"),
                                      sex = c("F", "M"))))
point
truth
MSE(point = point,
    truth = truth)

## calculations with MSE
m <- MSE(point = point,
         truth = truth)
class(m) # same as 'truth'
## mean
collapseDimension(m,
                  dimension = "sex",
                  weights = 1)
## sum
m <- as(m, "Counts")
collapseDimension(m,
                  dimension = "sex")
}
