% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemographicAccount-methods.R,
%   R/DemographicArray-methods.R, R/MetaData-methods.R
\docType{methods}
\name{names,DemographicAccount-method}
\alias{names,DemographicAccount-method}
\alias{names-methods}
\alias{names,DemographicArray-method}
\alias{names<-,DemographicArray-method}
\alias{names,MetaData-method}
\alias{names<-,MetaData-method}
\title{Get or set dimension names}
\usage{
\S4method{names}{DemographicAccount}(x)

\S4method{names}{DemographicArray}(x)

\S4method{names}{DemographicArray}(x) <- value

\S4method{names}{MetaData}(x)

\S4method{names}{MetaData}(x) <- value
}
\arguments{
\item{x}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{value}{Character vector.}
}
\value{
A character vector.  The replacement method returns an object of
class \code{\linkS4class{DemographicArray}}.
}
\description{
Query or change the dimension names of a \code{\linkS4class{DemographicArray}}
object.
}
\examples{
library(demdata)
popn <- Counts(VAPopn)
names(popn)
names(popn) <- toupper(names(popn))
}
\seealso{
As described in \code{\link{dimtypes}}, some dimensions must come
in pairs, and there are special rules governing the names of these
dimensions.
}
\author{
John Bryant \email{demographic.packages@gmail.com}
}
