% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Component-generator.R
\docType{methods}
\name{net-pool-generators}
\alias{net-pool-generators}
\alias{Net}
\alias{Pool}
\alias{Net,Counts-method}
\alias{Pool,Counts-method}
\title{Create object of class "Net" and "Pool".}
\usage{
Net(object, between)

Pool(object, direction, between)

\S4method{Net}{Counts}(object, between)

\S4method{Pool}{Counts}(object, direction, between)
}
\arguments{
\item{object}{An object of class \linkS4class{Counts}.}

\item{between}{Names or indices of dimensions of \code{object}}

\item{direction}{Name or index of dimension of \code{object}}
}
\description{
Object of class \code{\linkS4class{Net}} and \code{\linkS4class{Pool}}
are used to describe net flows, or 'in' and 'out' flows, typically
to be supplied to function \link{Movements}.
}
\details{
The 'between' dimensions \code{object} are the dimensions between which
flows occur.  For instance, with internal migration, the 'between'
dimension will be something like region.  There can be more than one
'between' dimension, e.g. region and employment status.

The 'direction' dimension, which is used only with \code{Pool} identifies
'in' flows and 'out' flows.  The dimension must have length 2 and must
have categories \code{"in"} and \code{"out"} or \code{"ins"} and \code{"outs"}
(case insensitive).  \code{Pool} converts the labels to \code{"In"}
and \code{"Out"}.
}
