% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{nIteration}
\alias{nIteration}
\alias{nIteration,DemographicArray-method}
\title{Get the number of iterations  in a demographic array.}
\usage{
nIteration(object)

\S4method{nIteration}{DemographicArray}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}
}
\value{
An integer.
}
\description{
If an object of class \code{\linkS4class{DemographicArray}} has a dimension
with \code{\link{dimtype}} \code{"iteration"}, return the lenth of that
dimension; if not, raise an error.
}
\examples{
library(demdata)
x <- Counts(array(replicate(n = 5, rnorm(2)),
                  dim = c(2, 5),
                  dimnames = list(sex = c("Female", "Male"),
                      iteration = 1:5)))
nIteration(x)

x <- Counts(VAPopn)
## no dimension with dimtype "iteration"
\dontrun{nIteration(x)}
}
\author{
John Bryant \email{demographic.packages@gmail.com}
}
\seealso{
\code{\link{dimtypes}}, \code{\link{collapseIterations}}
}
\keyword{manip}

