% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\name{pairToState}
\alias{pairToState}
\alias{pairToState,DemographicArray-method}
\title{Convert dimensions with paired dimtypes to state dimtypes}
\usage{
pairToState(object)

\S4method{pairToState}{DemographicArray}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}
}
\value{
An object with the same class as \code{object},
but with altered dimtypes and names, if any paired dimtypes were
present in \code{object}.
}
\description{
\code{\link{dimtypes}} \code{"origin"} and \code{"destination"},
and \code{\link{dimtypes}} \code{"parent"} and \code{"child"},
always come in pairs. If one of these dimensions is collapsed,
the remaining dimension is converted to dimtype \code{"state"},
and the \code{"_orig"}, \code{"_dest"}, \code{"_parent"}
or \code{"_child"} suffix in its name disappears.
Normally this behaviour is useful, but sometimes it is not.
\code{pairToState} converts all dimensions
in \code{object} with \code{\link{dimtype}} \code{"origin"},
\code{"destination"}, \code{"parent"}, or \code{"child"}
to \code{\link{dimtype}} \code{"state"}. To conform to the rule
that paired dimensions, and only paired dimensions, have
suffixes code{"_orig"}, \code{"_dest"}, \code{"_parent"}
or \code{"_child"}, the names of transformed dimensions
are changed, with the \code{"_"} characters in the
suffixes being changed to \code{"."}.
}
\examples{
x <- Counts(array(1:9,
                  dim = c(3, 3),
                  dimnames = list(region_orig = c("A", "B", "C"),
                                  region_dest = c("A", "B", "C"))))
x
pairToState(x)
}
\seealso{
For descriptions of dimtypes and the rules that govern
them, see \code{\link{dimtypes}}.  Dimensions are collapsed
using function \code{\link{collapseDimension}}.
}
