% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\name{perturb}
\alias{perturb}
\alias{perturb,DemographicArray-method}
\title{Generate demographic array consisting of random values.}
\usage{
perturb(object, n = 1L, order = 2L, phi = 1, subtotals = NULL)

\S4method{perturb}{DemographicArray}(object, n = 1L, order = 2L, phi = 1, subtotals = NULL)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{n}{Number of iterations to generate.}

\item{order}{Order of interations used in log-linear models.}

\item{phi}{A positive number governing the amount of random noise: higher
values imply more variation.}

\item{subtotals}{Not implemented yet}
}
\value{
An object of the same class as \code{object}.  If \code{n > 1}, the
return value has a dimension with dimtype \code{"Iterations"}; otherwise it
does not.
}
\description{
If \code{object} has a dimension with \code{\link{dimtype}}
\code{"Iterations"}, then \code{perturb} returns \code{n} randomly-selected
iterations.  Otherwise, it fits a log-linear model to the data and uses this
to randomly generate \code{n} iterations.  The log-linear model by default
contains 2-way interactions; higher or lower values can be specified via
\code{order}.
}
\details{
Arguments \code{order} and \code{phi} only have an effect when a log-linear
model is fitted - in other words, when \code{object} does not have dimtype
\code{"Iterations"}.

\code{object} can contain missing values, though if a log-linear model is
fitted, there must be sufficient observations for a model with order-way
interations.
}
\examples{
library(demdata)
x <- Values(VADeaths2)
perturb(x)
perturb(x, phi = 2) ## greater variance
perturb(x, order = 1) ## log-linear model only includes main effects
perturb(x, n = 2)
x10 <- perturb(x, n = 10)
perturb(x10, n = 2)  ## iterations selected, rather than generated
}
\seealso{
The log-linear model is fitted using \code{\link{loglm}}.
}
