% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R, R/Values-methods.R
\docType{methods}
\name{plot-methods}
\alias{plot,Counts-method}
\alias{plot,Values-method}
\alias{plot-methods}
\title{Plot a demographic array}
\usage{
\S4method{plot}{Counts}(x, main = NULL, cex.main = 1.2,
  col.main = "black", font.main = 2, las = 1, ...)

\S4method{plot}{Values}(x, threshold = 20, main = NULL, cex.main = 1.2,
  col.main = "black", font.main = 2, las = 1, ...)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{main}{Title for plot.}

\item{cex.main, col.main, font.main}{The 'character expansion factor', colour,
and font to be used for \code{main}.}

\item{las}{Parameter governing orientation of axis labels.  One of \code{0},
\code{1}, \code{2}, or \code{3}.  See \code{\link[graphics]{par}}.}

\item{threshold}{An integer.  If the number of values being plotted per
category is less than or equal to \code{threshold} then a boxplot is used;
otherwise a stripchart is used.}

\item{\dots}{Other arguments, which are passed to the underlying plotting
functions.}
}
\description{
Methods for function \code{\link{plot}} which aim to show the main features
of a demographic array.  Different formats are used
for objects of class \code{\linkS4class{Counts}} and
\code{\linkS4class{Values}}.
}
\examples{
library(demdata)
popn <- Counts(VAPopn)
plot(popn)

rates <- Values(VADeaths2)
plot(rates)
}
\seealso{
\code{plot} calls functions \code{\link{barchart}},
\code{\link{boxplot}}, and \code{\link{stripchart}}
}

