% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{prop.table}
\alias{prop.table}
\alias{prop.table,DemographicArray,ANY-method}
\alias{prop.table,DemographicArray,character-method}
\title{Calculate proportions within dimensions of a demographic array.}
\usage{
prop.table(x, margin = NULL)

\S4method{prop.table}{DemographicArray,ANY}(x, margin = NULL)

\S4method{prop.table}{DemographicArray,character}(x, margin = NULL)
}
\arguments{
\item{x}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{margin}{Indices or names of dimensions within which proportions are to
be calculated.}
}
\value{
Object of class \code{\linkS4class{Values}}.
}
\description{
The is a method for function \code{\link[base]{prop.table}} from package
\pkg{base}.  Unlike the original function, dimension names can be used in
place of dimension indices.
}
\details{
If \code{x} has a dimension with \code{\link{dimtype}} \code{"Iterations"},
and that dimension is not explicitly included in \code{margin}, then it is
added automatically.
}
\examples{
library(demdata)
x <- Counts(VAPopn)
prop.table(x, margin = 2:4)
## the equivalent, using dimension names
prop.table(x, margin = c("sex", "residence", "color"))
}
