% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\docType{methods}
\name{reallocateToEndAges}
\alias{reallocateToEndAges}
\alias{reallocateToEndAges,Counts,ANY,ANY,missing-method}
\alias{reallocateToEndAges,Counts,ANY,ANY,ANY-method}
\alias{reallocateToEndAges,Values,ANY,ANY,missing-method}
\alias{reallocateToEndAges,Values,ANY,ANY,Counts-method}
\title{Reallocate counts or values to youngest and oldest age groups.}
\usage{
reallocateToEndAges(object, min = 15, max = 50, weights, ...)

\S4method{reallocateToEndAges}{Counts,ANY,ANY,missing}(object, min = 15,
  max = 50, weights, ...)

\S4method{reallocateToEndAges}{Counts,ANY,ANY,ANY}(object, min = 15,
  max = 50, weights, ...)

\S4method{reallocateToEndAges}{Values,ANY,ANY,missing}(object, min = 15,
  max = 50, weights, ...)

\S4method{reallocateToEndAges}{Values,ANY,ANY,Counts}(object, min = 15,
  max = 50, weights, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}} or (less commonly)
\code{\linkS4class{Values}}.}

\item{min}{The lower limit of the youngest age group, after restriction of
the age range.}

\item{max}{The upper limit of the oldest age group, after restriction
of the age range.}

\item{weights}{An object of class \code{\linkS4class{Counts}}.}

\item{\dots}{Not currently used.}
}
\value{
A \code{\linkS4class{Counts}} object in which all
age groups lie within the range (\code{min}, \code{max}).
}
\description{
With fertility data, it is common to confine age-specific fertility rates
to a limited range of age-gropus, eg ages 12 to 49, or ages 15-19 to 45-49.
Births outside these age groups are reallocated to the bottom' and top
age groups.  The procedure preserves the total number of birth and total
fertility, and has only a minor effect
on the age profile for fertility, provided that the number of births that are
reallocated is small.
}
\details{
\code{reallocateToEndAges} reallocates counts or rates.  There is no
requirement' that the counts or rates relate to fertility,
though this is by far the
most common scenario.

If \code{min} is set to a value that is less than the lower limit
of the youngest age group in \code{object}, then the return value
from \code{reallocateToEndAges} has the same youngest age group as
\code{code}.  Similarly, if \code{max} is set to a value higher
than the upper limit of the oldest age group,
then \code{reallocateToEndAges} leaves the oldest age group untouched.

A \code{weights} argument is prohibited if \code{object} has class
\code{\linkS4class{Counts}}, and required if it has class
\code{\linkS4class{Values}}.
}
\examples{
## make some fake data on births
births <- Counts(array(10,
                       dim = c(45, 2),
                       dimnames = list(age = 10:54, region = c("A", "B"))),
                 dimscales = c(age = "Intervals"))

## default ages
reallocateToEndAges(births)
## non-default ages
reallocateToEndAges(births, min = 12, max = 45)

## make some fake data on birth rates
birth.rates <- Values(array(0.1,
                      dim = c(45, 2),
                      dimnames = list(age = 10:54, region = c("A", "B"))),
                      dimscales = c(age = "Intervals"))
women <- Counts(array(100,
                      dim = c(45, 2),
                      dimnames = list(age = 10:54, region = c("A", "B"))),
                dimscales = c(age = "Intervals"))
## with rates, need to supply weights
reallocateToEndAges(birth.rates, weights = women)                       
}
\seealso{
\code{\link{collapseIntervals}} collapses age intervals
without reallocating values upwards or downwards.
}
