% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\docType{methods}
\name{redistributeCategory}
\alias{redistributeCategory}
\alias{redistributeCategory,Counts-method}
\title{Redistribute values from one or more categories among the remaining categories.}
\usage{
redistributeCategory(counts, dimension, category, means = FALSE,
  epsilon = 0, n = NULL)

\S4method{redistributeCategory}{Counts}(counts, dimension, category,
  means = FALSE, epsilon = 0, n = NULL)
}
\arguments{
\item{counts}{Object of class \code{\linkS4class{Counts}} consisting of
non-negative integers, or a single non-negative integer.}

\item{dimension}{Names or indices for dimensions whose categories are to be
reallocated. If the dimensions have \code{\link{dimtypes}}
\code{"origin"} and \code{"destination"} or \code{"parent"} and
\code{"child"}, then the base names may be used.}

\item{category}{Names or indices of the categories to be reallocated.}

\item{means}{If \code{FALSE} (the default) the allocation is done randomly;
if \code{TRUE}, values are set equal to their means.}

\item{epsilon}{Small quantity added to counts for remaining categories.}

\item{n}{Number of iterations to generate.  Ignored if \code{counts} or
\code{weights} has a dimension with \code{\link{dimtype}}
\code{"iteration"}.}
}
\value{
A \code{\linkS4class{Counts}} object without the categories
specified by \code{categories}.
}
\description{
Take counts from one or more categories (eg an "Other" category") and allocate
them the remaining categories, in proportion to the current size of those categories.
}
\details{
When \code{means} is \code{FALSE} (the default) the allocation is done
randomly, via multinomial draws.  When \code{means} is \code{FALSE} the
allocation is done deterministically, and exactly matches the existing
distribution.

Argument \code{epsilon} provides a quick and dirty way of using the function
with sparse data, when all of the remaining categories may be zero (so that
the weights are undefined.)  The quantity \code{epsilon} is added to all
counts before calculating.  With sparse data a better approach would be to
fit a model and then redistribute using the fitted values, using function
\code{\link{redistribute}}.
}

