% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R, R/Values-methods.R
\docType{methods}
\name{redistributeToEndAges}
\alias{redistributeToEndAges}
\alias{redistributeToEndAges,Counts,ANY,ANY,ANY-method}
\alias{redistributeToEndAges,Counts,ANY,ANY,missing-method}
\alias{redistributeToEndAges,Values,ANY,ANY,Counts-method}
\alias{redistributeToEndAges,Values,ANY,ANY,missing-method}
\title{Redistribute counts or values into youngest and oldest age groups.}
\usage{
redistributeToEndAges(object, min = 15, max = 50, weights, ...)

\S4method{redistributeToEndAges}{Counts,ANY,ANY,missing}(object, min = 15,
  max = 50, weights, ...)

\S4method{redistributeToEndAges}{Counts,ANY,ANY,ANY}(object, min = 15,
  max = 50, weights, ...)

\S4method{redistributeToEndAges}{Values,ANY,ANY,missing}(object, min = 15,
  max = 50, weights, ...)

\S4method{redistributeToEndAges}{Values,ANY,ANY,Counts}(object, min = 15,
  max = 50, weights, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}} or (less commonly)
\code{\linkS4class{Values}}.}

\item{min}{The lower limit of the youngest age group, after reassignment.}

\item{max}{The upper limit of the oldest age group, after reassignment.}

\item{weights}{An object of class \code{\linkS4class{Counts}}.}

\item{\dots}{Not currently used.}
}
\value{
A \code{\linkS4class{Counts}} object without the categories
specified by \code{categories}.
}
\description{
With fertility data, it is common to confine age-specific fertility rates
to a limited range, eg ages 12-49, or 15-19 to 45-49.  Births outside these
ages are conventionally reassigned to the bottom and top ages.  The procedure
preserves the total number of birth and total fertility, and has only a minor effect
on the age profile for fertility, provided that the number of births that are
reassigned is small.
}
\details{
\code{redistributeToEndAges} reassigns counts or rates.  There is no requirement
that the counts or rates relate to fertility, though this is by far the
most common scenario.

A \code{weights} argument is prohibited if \code{object} has class
\code{\linkS4class{Counts}}, and required if it has class
\code{\linkS4class{Values}}.
}
\examples{
## make some fake data on births
births <- Counts(array(10,
                       dim = c(45, 2),
                       dimnames = list(age = 10:54, region = c("A", "B"))),
                 dimscales = c(age = "Intervals"))

## default ages
redistributeToEndAges(births)
## non-default ages
redistributeToEndAges(births, min = 12, max = 45)

## make some fake data on birth rates
birth.rates <- Values(array(0.1,
                      dim = c(45, 2),
                      dimnames = list(age = 10:54, region = c("A", "B"))),
                      dimscales = c(age = "Intervals"))
women <- Counts(array(100,
                      dim = c(45, 2),
                      dimnames = list(age = 10:54, region = c("A", "B"))),
                dimscales = c(age = "Intervals"))
## with rates, need to supply weights
redistributeToEndAges(birth.rates, weights = women)                       
}
\seealso{
\code{\link{collapseIntervals}} collapses age intervals
without reassigning values upwards or downwards.
}

