% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R
\name{resetDiag}
\alias{resetDiag}
\alias{resetDiag,Counts,ANY,ANY-method}
\alias{resetDiag,Counts,ANY,missing-method}
\alias{resetDiag,Counts,ANY,NULL-method}
\alias{resetDiag,Values,ANY,ANY-method}
\alias{resetDiag,Values,ANY,NULL-method}
\alias{resetDiag,Values,ANY,missing-method}
\title{Reset values where origin equals destination.}
\usage{
resetDiag(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Counts,ANY,ANY}(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Counts,ANY,missing}(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Counts,ANY,`NULL`}(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Values,ANY,ANY}(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Values,ANY,`NULL`}(object, base = NULL, reset = NULL)

\S4method{resetDiag}{Values,ANY,missing}(object, base = NULL, reset = NULL)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}

\item{base}{Character vector giving the "base names" for the
origin-destination pairs.  If \code{base} is omitted, the base names from
all dimensions with \code{\link{dimtype}} \code{"origin"} and
\code{"destination"} are used.}

\item{reset}{The new value to be used on the diagonal.}
}
\value{
The original object, with the objects on the diagonal changed.
}
\description{
Given a \code{\linkS4class{DemographicArray}} with \code{\link{dimtypes}}
\code{"origin"} and \code{"destination"}, set cells on the "diagonal",
ie where the origin equals the destination, to a value such as \code{0}
or \code{NA}.
}
\details{
The origin and destination dimensions are specified via the \code{base}
argument.  \code{base} is the "base name", that is, the first part
of the dimension name.  If is an origin dimension called
\code{"region_orig"}, for instance, and a destination called
\code{"region_dest"}, then the base name is \code{"region"}.
If no \code{base} argument is supplied, then all origin and
destination dimensions in the object are used.

If no \code{reset} argument is supplied, then it defaults to
\code{0} in the case of \code{\linkS4class{Counts}} objects and
\code{NA} in the case of \code{\linkS4class{Values}} objects.
}
\examples{
x <- Counts(array(1:4,
                  dim = c(2, 2),
                  dimnames = list(reg_orig = c("A", "B"),
                                  reg_dest = c("A", "B"))))
resetDiag(x)
resetDiag(x, base = "reg")
resetDiag(x, reset = NA)

## with Values object, reset defaults to NA, not 0
x.val <- as(x, "Values")
resetDiag(x.val)
}
\seealso{
\code{\link{collapseOrigDest}} to collapse away
origin and destination dimensions.  \code{\link{alignPair}}
puts origin and destination categories in the same order.
}
