% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R,
%   R/MetaData-methods.R
\name{resetIterations}
\alias{resetIterations}
\alias{resetIterations,DemographicArray-method}
\alias{resetIterations,MetaData-method}
\title{Reset iteration labels.}
\usage{
resetIterations(object)

\S4method{resetIterations}{DemographicArray}(object)

\S4method{resetIterations}{MetaData}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}
}
\description{
Reset the labels used by a dimension with \code{\link{dimtype}}
\code{"iteration"} to be \code{1:nIteration(object)}.
}
\examples{
x <- ValuesOne(rnorm(10), labels = seq(2, 20, 2), name = "iteration")
resetIterations(x)
}
