% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\name{rotateAgeTime}
\alias{rotateAgeTime}
\alias{rotateAgeTime,DemographicArray-method}
\title{Rotate the age-time plan used by a demographic array}
\usage{
rotateAgeTime(object, to = NULL, name = NULL)

\S4method{rotateAgeTime}{DemographicArray}(object, to = NULL, name = NULL)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}}.}

\item{to}{The dimtypes of \code{object} after the age-time
plan has been rotated.}

\item{name}{The name of the new age, time, or cohort dimension.}
}
\value{
An object with the same class as \code{object}.
}
\description{
Convert a \code{\linkS4class{DemographicArray}} between
age-time, age-chort, and time-cohort formats.
}
\details{
If the array has dimensions with two out of the three
\code{\link{dimtypes}} \code{"age"}, \code{"time"},
\code{"cohort"}, then replace one of these dimensions
with a new dimension with the missing dimtype, and
rearrange the array accordingly. If,
for instance, an array has dimensions with dimtypes
\code{"age"} and \code{"time"}, then replace either
the age or time dimension with a cohort dimension,
and rearrange cells within the array accordingly.

\code{object} must have a regular age-time plan
(see \code{\link{hasRegularAgeTime}}. Negative ages
are not allowed, and cohorts must not start
later than the last time point or period. \code{object}
must also not contain any "open" age intervals,
time intervals, or cohorts - ie intervals that include
positive or negative infinity.

The argument \code{to} gives the dimensions used
in the new age-time plan. It can be one of
\code{"age-time"}, \code{"time-age"},
\code{"age-cohort"}, \code{"cohort-age"},
\code{"time-chort"}, or \code{"cohort-time"}.
The abbreviations \code{"at"}, \code{"ta"},
\code{"ac"}, \code{"ca"}, \code{"tc"}, or \code{"ct"}
can also be used. (For production code, however,
the full versions are better, since they
are more self-documenting.) \code{to} is
not case-sensitive, and only enough characters to
uniquely identify the option need to be supplied.

By default the new dimension is named after the
corresponding dimtype: for instance, if the new
dimension has dimtype \code{"cohort"}, then the
new dimension is named \code{"cohort"}.  However,
an alternative name can be provided via the
\code{name} argument.
}
\examples{
x <- Counts(array(1:6,
                  dim = c(3, 2),
                  dimnames = list(age = c("0-4", "5-9", "10-14"),
                                  time = c(2000, 2005))))
rotateAgeTime(x, to = "age-cohort")
}
\seealso{
\code{\link{dimtypes}}
}
