% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R,
%   R/Values-methods.R, R/array-methods.R, R/numeric-methods.R
\name{round3}
\alias{round3}
\alias{round3,Counts-method}
\alias{round3,Values-method}
\alias{round3,array-method}
\alias{round3,numeric-method}
\title{Randomly round a dataset to base 3.}
\usage{
round3(object)

\S4method{round3}{Counts}(object)

\S4method{round3}{Values}(object)

\S4method{round3}{array}(object)

\S4method{round3}{numeric}(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{DemographicArray}},
an ordinary array, or a numeric vector.}
}
\value{
An object with the same class as \code{object}.
}
\description{
Confidentialise counts data by randomly rounding it to multiples of three.
}
\details{
Randomly rounding to base three is a popular confidentialisation technique
at statistical agencies.  It only works with data consisting entirely of
integers, though negative integers are allows.

A value is rounded as follows:
\itemize{
  \item{If the value is  divisible by 3, leave it unchanged.}
  \item{If there is a remainder of 1 after dividing by 3, round down (ie subtract 1)
with probability 2/3, and round up (ie add 2) with probability 1/3.}
  \item{If there is a remainder of 2 after dividing by 3, round down (ie subtract 2)
with probability 1/3, and round up (ie add 1) with probability 2/3.}
}
Missing values are permitted, and are left unchanged.
}
\examples{
x <- CountsOne(0:10, labels = letters[1:11], name = "region")
round3(x)
round3(x)
}
