% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{splice}
\alias{splice}
\title{Splice concordances.}
\usage{
splice(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{Concordance}}.}
}
\value{
An object of class \code{\linkS4class{Concordance}}.
}
\description{
Combine two or more concordances to form a new concordance, via
a \code{\link[base]{rbind}}-like operation.
}
\details{
Concordances that are to be spliced together must have classifications with
the same names.  The result of the splicing must also be a many-to-one or
one-to-one relationship between classifications.  Duplicate rows are
deleted.
}
\examples{
x <- cbind(c1 = c("a", "b", "c"), c2 = c("q", "r", "r"))
x <- Concordance(x)
y <- cbind(c1 = c("e", "f"), c2 = c("s", "t"))
y <- Concordance(y)
x
y
splice(x, y)

x <- cbind(c1 = c("a", "b", "c"), c2 = c("q", "r", "r"))
x <- Concordance(x)
y <- cbind(c2 = c("r", "s"), c1 = c("c", "d"))
y <- Concordance(y)
x
y  ## columns in different order, and rows overlap with 'x'
splice(x, y)

}
\seealso{
\code{\link{chain}} combines concordances via a
\code{\link[base]{merge}}-like operation.
}
