% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{Statistical-Functions}
\alias{Statistical-Functions}
\alias{Summary,DemographicArrayOrNumeric-method}
\alias{mad,DemographicArray-method}
\alias{mean,DemographicArray-method}
\alias{median,DemographicArray-method}
\title{Existing statistical functions.}
\usage{
\S4method{Summary}{DemographicArrayOrNumeric}(x, ..., na.rm = FALSE)

\S4method{mad}{DemographicArray}(x, center = median(x), constant = 1.4826,
  na.rm = FALSE, low = FALSE, high = FALSE)

\S4method{mean}{DemographicArray}(x, ...)

\S4method{median}{DemographicArray}(x, na.rm = FALSE)
}
\arguments{
\item{x}{Typically an object of class \code{\linkS4class{DemographicArray}}.}

\item{na.rm}{Logical. Whether to remove missing values before calculating.}

\item{center}{Function used to calculate central value.  Defaults
to \code{\link[stats]{median}}.}

\item{constant}{See documentation for \code{\link[stats]{mad}}.}

\item{low}{See documentation for \code{\link[stats]{mad}}.}

\item{high}{See documentation for \code{\link[stats]{mad}}.}

\item{\dots}{Other arguments.}
}
\value{
Numeric, if \code{x} does not contain an \code{"iteration"}
dimension, and \code{\link{Counts}} or \code{\link{Values}} if it
does.
}
\description{
Existing statistical functions give the same result when used with a
\code{\linkS4class{DemographicArray}} array as they would with an
ordinary \code{\link[base]{array}} - except when the demographic array
has a dimension with \code{\link{dimtype}} \code{"iteration"} or
\code{"quantile"}.  When there is an \code{"iteration"} dimension,
the calculations are repeated once for each iteration, so that the
return value is also a demographic array with an \code{"iteration"}
dimension.
}
\details{
When called on an ordinary array, the functions all return a single
number, except for \code{range}, which returns 2.  When called on
a demographic array with an \code{"iteration"} dimension, \code{range}
returns a demographic array with a dimension called \code{"range"},
in addition to the \code{"iteration"} dimension.

When a demographic array has a dimension with \code{\link{dimtype}}
\code{"quantile"}, applying statistical functions typically does
not make sense, and an error is raised.  It is generally better to work
with the original array, with the complete set of iterations, and to
reduce to quantiles once the calculations are complete.
}
\examples{
## no iteration dimension
x1 <- Counts(array(1:4,
                   dim = 4,
                   dimnames = list(region = c("A", "B", "C", "D"))))
## with iteration dimension
x2 <- Counts(array(c(1:4, 0:3, 2:5),
                   dim = c(4, 3),
                   dimnames = list(region = c("A", "B", "C", "D"),
                                   iteration = 1:3)))
x1
x2
sum(x1)
sum(x2)
range(x1)
range(x2)
}
\seealso{
Iteration and quantile dimensions are described in the
documentation for \code{\link{dimtypes}}.  To remove iteration
dimensions (by summarising across them) use
\code{\link{collapseIterations}}.
}

