% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{subarray}
\alias{subarray}
\alias{subarray,DemographicArray-method}
\title{Subset a demographic array.}
\usage{
subarray(object, ...)

\S4method{subarray}{DemographicArray}(object, subarray, drop = TRUE)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{subarray}{Logical expression picking out a subarray within
\code{object}.}

\item{drop}{Logical, specifying whether dimensions of length 1 should be
retained in the result.  Defaults to \code{TRUE}.}

\item{\dots}{Further arguments to be passed to or from other methods.}
}
\value{
Typically an object of class \code{\linkS4class{DemographicArray}}, but
maybe a vector in some cases when \code{drop} is \code{TRUE}.
}
\description{
Use a logical expression to select a subarray from within an object of class
\code{\linkS4class{DemographicArray}}.
}
\details{
Selections involving \code{\link{dimscales}} \code{Points} or
\code{Intervals} can be made in the same way as selections involving other
dimscales, by specifying labels (in quotes.) However, they can also be made
using inequality operators \code{<}, \code{<=}, \code{>}, and \code{>=}.  An
intervals is selected by an inequality operator if the operator takes in the
entire interval.  For instance, interval \code{(a, a+n)} on dimension
\code{age} is selected by the expression \code{age > x} if \code{x <= a}.
The interval is not selected by \code{a < x}, even if \code{x < a + n}.
Expression \code{age > x} has the same effect as \code{age >= x}, based on
the assumption that no one is age \emph{exactly} \code{a} (and similarly for
time).

\code{subarray} sometimes modifies the names, \code{\link{dimtypes}} or
\code{\link{dimscales}} of the return value so that they conform to the
rules governing dimtypes and dimscales.  For instance, when the return value
has only one dimension from an origin-destination pair, \code{subarray} will
change the dimtype of that dimension to \code{"state"}, and will remove the
\code{"_orig"} or \code{"_dest"} suffix from its name.  See below for a
second example where \code{subarray} changes the dimtypes and dimscales of
an \code{"age"} dimension.
}
\section{Warning}{
 The documentation for function \link[base]{subset} (the
inspiration for \code{subarray}) warns that the non-standard evaluation of
the \code{subset} argument can have unanticipated consequences.  It
recommends that, for programming, subsetting be done via standard functions
such as \code{[}.  The same advice applies to \code{subarray}: see below for
an example.  However, these sorts of problems normally only occur when using
\code{subarray} within another function.  In interactive use, or in simple
scripts to extract or manipulate a dataset, the greater transparency and
flexibility of \code{subarray} typically make it a better choice than
\code{[}.
}
\examples{
library(demdata)
rates <- Values(VADeaths2)
## selection using labels
subarray(rates, residence == "Rural")
subarray(rates, age == "50-54")
subarray(rates, age \%in\% c("50-54", "55-59"))

## selection using inequalities
subarray(rates, age > 60)
subarray(rates, age > 59)
subarray(rates, age >= 60)

## selection creates a gap in the age groups,
## so dimscales changes to "Categories",
## which means dimtypes must change to "state"
subarray(rates, age < 55 | age > 65)

## can select from multiple dimensions at once
subarray(rates, (age < 60) & (sex == "Male"))
## but only if the selection is 'rectangular'
## - ie results in an array
\dontrun{subarray(rates, (age > 60) | (sex == "Male"))}

mig <- Counts(nz.mig)
subarray(mig, island_orig == "South Island")

## an example of the problems encountered calling
## 'subarray' from within a function
f <- function(orig) {
    subarray(mig, island_orig = orig)
}
\dontrun{f()}
}
\seealso{
\code{\link[base]{subset}} provides similar functionality for
vectors, matrices, or data frames, though without the requirement that the
resulting subsets have all the combinations of values required for an array.
}

