% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/SubArrayIndices-methods.R
\docType{class}
\name{SubArrayIndices-class}
\alias{!,SubArrayIndices-method}
\alias{&,SubArrayIndices,SubArrayIndices-method}
\alias{SubArrayIndices-class}
\alias{|,SubArrayIndices,SubArrayIndices-method}
\title{Class used by function 'subarray'.}
\usage{
\S4method{&}{SubArrayIndices,SubArrayIndices}(e1, e2)

\S4method{|}{SubArrayIndices,SubArrayIndices}(e1, e2)

\S4method{!}{SubArrayIndices}(x)
}
\description{
For internal use only
}
\keyword{internal}

