% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{thinIterations}
\alias{thinIterations}
\alias{thinIterations,DemographicArray-method}
\title{Reduce the number of iterations in demographic array.}
\usage{
thinIterations(object, n)

\S4method{thinIterations}{DemographicArray}(object, n)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{n}{The number of iterations to retain.}
}
\description{
Given an object of class \code{\linkS4class{DemographicArray}} that has a
dimension with \code{\link{dimtype}} \code{"iteration"}, return a new
object containing a subset of the original iterations.
}
\details{
The \code{n} iterations to retain are chosen at random, and are numbered
\code{1}, \code{2}, ..., \code{n}.

If \code{object} does not have a dimension with dimtype \code{"iteration"},
\code{thinIterations} throws an error.
}
\examples{
x <- Counts(array(1:6,
                  dim = c(2, 3),
                  dimnames = list(sex = c("Female", "Male"), iteration = 1:3)))
thinIterations(x, n = 2)
}
\seealso{
Other functions that can be used to subset a dimension with
\code{\link{dimtype}} \code{"iteration"} are \code{\link{subarray}},
\code{\link{slab}}, and \code{\link{[}}.  To remove an \code{"iteration"}
dimension entirely, use \code{\link{collapseIterations}}.
}

