% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Concordance-methods.R
\docType{methods}
\name{translate}
\alias{translate}
\alias{translate,ANY,ManyToOne,missing-method}
\alias{translate,ANY,ManyToOne,ANY-method}
\alias{translate,ANY,OneToOne,ANY-method}
\alias{translate,ANY,OneToOne,NULL-method}
\alias{translate,ANY,OneToOne,missing-method}
\alias{translate,ANY,data.frame,ANY-method}
\alias{translate,ANY,matrix,ANY-method}
\title{Translate between classifications.}
\usage{
translate(object, concordance, to = NULL, ...)

\S4method{translate}{ANY,ManyToOne,missing}(object, concordance)

\S4method{translate}{ANY,ManyToOne,ANY}(object, concordance, to)

\S4method{translate}{ANY,OneToOne,ANY}(object, concordance, to = NULL)

\S4method{translate}{ANY,OneToOne,`NULL`}(object, concordance, to = NULL)

\S4method{translate}{ANY,OneToOne,missing}(object, concordance)

\S4method{translate}{ANY,data.frame,ANY}(object, concordance, to = NULL,
  ...)

\S4method{translate}{ANY,matrix,ANY}(object, concordance, to = NULL, ...)
}
\arguments{
\item{object}{A character vector.}

\item{concordance}{An object of class \code{\linkS4class{Concordance}}.}

\item{to}{Name of classification within \code{concordance} that codes from
\code{object} should be translated to.}

\item{\dots}{Not currently used.}
}
\value{
A character vector, the same length as \code{object}.
}
\description{
Translate between classifications, based on a concordance.
}
\details{
If \code{concordance} is \code{\linkS4class{ManyToOne}}, then
\code{object} must be the 'many' part, and argument \code{to} is not
required.  If \code{concordance} is \code{\linkS4class{OneToOne}}, and
\code{to} is not supplied, \code{translate} guess, based on the values
in \code{object}.
}
\examples{
conc <- data.frame(code = 1:3,
                   descriptor = c("North Island", "South Island",
                                   "Stewart Island"))
conc <- Concordance(conc)
conc
translate(c("1", "3", "1", "2"),
          concordance = conc)
translate(c("Stewart Island", "North Island"),
          concordance = conc)


codes(conc, classification = "code")
codes(conc, classification = "descriptor")

conc <- data.frame(sex1 = c("girl", "boy", "woman", "man"),
                   sex2 = c("female", "male"))
conc <- Concordance(conc)
conc
translate(c("man", "girl"), concordance = conc)
\dontrun{
translate(c("female", "male"), concordance = conc)
}
}
