% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{valueInInterval}
\alias{valueInInterval}
\alias{valueInInterval,DemographicArray,DemographicArray-method}
\alias{valueInInterval,numeric,DemographicArray-method}
\title{Test whether values lie within intervals}
\usage{
valueInInterval(value, interval, lower.inclusive = TRUE,
  upper.inclusive = TRUE)

\S4method{valueInInterval}{DemographicArray,DemographicArray}(value,
  interval, lower.inclusive = TRUE, upper.inclusive = TRUE)

\S4method{valueInInterval}{numeric,DemographicArray}(value, interval,
  lower.inclusive = TRUE, upper.inclusive = TRUE)
}
\arguments{
\item{value}{A \code{\linkS4class{DemographicArray}}, or a single
number.}

\item{interval}{A \code{\linkS4class{DemographicArray}},
with a quantile dimension of length 2.}

\item{lower.inclusive}{Logical. Whether the lower bounds
specified in \code{interval} are considered to lie within
the intervals.}

\item{upper.inclusive}{Logical. Whether the upper bounds
specified in \code{interval} are considered to lie
within the intervals. Defaults to \code{TRUE}.}
}
\value{
An array of logical values with the same dimensions
as \code{value}, or, if \code{value} is a number,
a single logical value.
}
\description{
Test whether the elements of \code{value} lie within the
lower and upper bounds specified by \code{interval}.
}
\details{
\code{interval} should have the same dimensions
and dimscales as \code{value}, but also have a dimension
with length 2 and \code{\link{dimtype}}
\code{"quantile"}. The quantile dimension specifies
the upper and lower bounds. See below for an example.

The \code{lower.inclusive} and \code{upper.inclusive}
arguments can be used to specify whether the intervals are
open or closed at the boundaries. The difference between
open and closed boundaries is typically neglible
when \code{value} consists of real numbers, but can
be large when \code{value} consists of integers, especially
if the integers are small.

\code{valueInInterval} is useful for simulation studies,
where it can be used to calculate the proportion of
credible intervals that contain the target values.

If \code{value} is a single number, then all dimensions of
\code{interval}, other than the quantile dimension, should
have length 1.

\code{valueInInterval} is stricter about the compatibility
of its arguments than most functions in \code{dembase}.
Although it reorders dimensions and categories
in \code{value} and \code{interval}, it does not collapse
or expand dimensions, or drop any levels.
}
\examples{
value <- ValuesOne(c(0.3, 0.5),
                   labels = c("Female", "Male"),
                   name = "sex")
interval <- Values(array(c(-0.1, 0.4, 0.2, 0.7),
                         dim = c(2, 2),
                         dimnames = list(sex = c("Female", "Male"),
                                         quantile = c("2.5\%", "97.5\%"))))
valueInInterval(value = value,
                interval = interval)

value <- 5L
interval <- ValuesOne(c(5L, 8L),
                      labels = c("5\%", "95\%"),
                      name = "quantile")
valueInInterval(value = value,
                interval = interval)
valueInInterval(value = value,
                interval = interval,
                lower.inclusive = FALSE)
}
\seealso{
Function \code{\link{credibleInterval}} is a convenience
function for creating \code{interval} objects of the right form
for \code{valueInInterval}. Function \code{\link{MSE}},
is another function that may be useful in simulation studies.
}
