% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{yearToPeriod}
\alias{yearToPeriod}
\title{Convert labels for single-year time periods to labels for multi-year
time periods}
\usage{
yearToPeriod(year, breaks, labelStart = TRUE, firstOpen = FALSE,
  lastOpen = FALSE)
}
\arguments{
\item{year}{A vector of year labels. A numeric vector, or a vector
than can be coerced to numeric.}

\item{breaks}{Numeric. A vector of breaks, specifying
points dividing periods.}

\item{labelStart}{Logical. Whether the single-year periods
described by \code{year} are labelled according to the year
at the start of the period or the year at the end.  Defaults
to \code{TRUE}.}

\item{firstOpen}{Logical. Whether the first period created by
\code{yearToPeriod} should be "open", i.e. has no lower bound.
Defaults to \code{FALSE}.}

\item{lastOpen}{Logical. Whether the last period created by
\code{yearToPeriod} is "open",
i.e. has no upper bound. Defaults to \code{FALSE}.}
}
\value{
A factor, the same length as \code{year}.
}
\description{
Convert a vector made of values such as \code{"2004", "2001", "2009"}
into a vector of values such as \code{"2000-2005", "2000-2005",
"2005-2010"}.  Period labels constructed using \code{yearToPeriod}
are  periods are formatted in the way expected by
functions such as \code{\link{Counts}} and
\code{\link{Values}}.
}
\details{
If the \code{year} argument is a factor, \code{yearToPeriod}
will coerce it to a character vector before trying to
coerce it to numeric. See below for an example.
}
\examples{
year <- c(2001, 2023, 2000, 2005, 2014, 2013, 2029)
## 5-year periods, 2000-2005, 2005-2010, ..., 2025-2030
yearToPeriod(year, breaks = seq(2000, 2030, 5))
## 10-year periods, 2000-2005, 2005-2010, ..., 2025-2030
yearToPeriod(year, breaks = seq(2000, 2030, 10))
## single-year periods specified by 'year' are labelled
## according to time at start
year <- 2000:2009
breaks <- c(1995, 2000, 2005, 2010)
yearToPeriod(year = year, breaks = breaks)
## single-year periods specified by 'year' are labelled
according to time at end
year <- 2000:2009
breaks <- c(1995, 2000, 2005, 2010)
yearToPeriod(year = year, breaks = breaks, labelStart = FALSE)
## first period open 
yearToPeriod(year, breaks = seq(2010, 2030, 5), firstOpen = TRUE)
}
\seealso{
\code{\link{ageToAgeGroup}}
}
