% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{datesToAge}
\alias{datesToAge}
\title{Calculate ages from dates.}
\usage{
datesToAge(date, dob)
}
\arguments{
\item{date}{A vector of class \code{\link[base]{Date}}.  All elements of
\code{date} must be equal to or greater than the corresponding elements
of \code{dob}.}

\item{dob}{A vector of class \code{\link[base]{Date}}.}
}
\value{
An vector of integers.
}
\description{
Calculates ages in completed years from (i) a vector of dates, and (ii)
a vector of dates of birth.
}
\details{
If \code{date} and \code{dob} are different lengths, the shorter
vector is recycled.
}
\examples{
dates <- as.Date(c("2005-07-05", "2010-07-05"))
dobs <- as.Date(c("2005-06-30", "2006-07-01"))
datesToAge(date = dates, dob = dobs)

## 'date' must be later than 'dob'
\dontrun{
datesToAge(date = as.Date("2010-01-01"), dob = as.Date("2000-01-01"))
}
}
\seealso{
Vectors of class \code{\link{Date}} can be created with
function \code{\link{as.Date}}.
}
