% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\name{exposure}
\alias{exposure}
\alias{exposure,Counts-method}
\title{Calculate exposure from population counts.}
\usage{
exposure(object, triangles = FALSE, openTriangles = c("weighted", "standard"))

\S4method{exposure}{Counts}(object, triangles = FALSE, openTriangles = c("weighted", "standard"))
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}}.}

\item{triangles}{Logical.  If \code{TRUE}, the return value
includes a dimension with \code{\link{dimtype}} code{"triangle"}.
Defaults to \code{FALSE}.}

\item{openTriangles}{\code{"weighted"} (the default) or \code{"standard"}.}
}
\value{
An object of class \code{\linkS4class{Counts}}.
}
\description{
\code{object} typically has a dimension with \code{\link{dimtype}}
\code{time} and \code{dimscale} \code{"Points"}, and a dimension
with dimtype \code{"age"} and dimscale \code{"Intervals"}.
The complete set of choices for age-time-cohort dimensions and
dimscales is
\tabular{ccc}{
  \code{time} \tab \code{age} \tab \code{cohort} \cr
  \code{Points} \tab \code{Intervals} \tab <none> \cr
  \code{Points} \tab <none> \tab <none> \cr
  <none> \tab \code{Points} \tab <none> \cr
  \code{Points} \tab <none> \tab \code{Intervals} \cr
  <none> \tab \code{Points} \tab \code{Intervals}
}
}
\details{
If \code{triangles} is \code{TRUE}, then exposure is calculated
separately for lower and upper Lexis triangles.  \code{triangles}
can only by \code{TRUE} if  \code{object} has dimensions with
\code{\link{dimtype}} \code{"time"} and \code{"age"}.

Argument \code{openTriangles} governs the way that the exposures
for the final age group are calculated when the final age group is
open (ie has no upper age limit) and \code{triangles} is \code{TRUE}.
The \emph{sum} of the exposures for the upper and lower triangles is unaffected.
When \code{openTriangles} is \code{"weighted"}, exposure for the upper triangle
is twice as large as exposure for the lower triangle (reflecting the
fact that the upper triangle includes multiple cohorts).
When \code{openTriangles} is \code{"standard"}
exposure is calculated the same way as it is for other age groups.
}
\examples{
## prepare data
library(demdata)
popn <- demdata::nz.popn.reg
popn <- Counts(popn, dimscales = c(year = "Points"))
popn <- subarray(popn, year \%in\% c(1996, 2001, 2006, 2011, 2016))
popn <- collapseDimension(popn, dimension = "region")

## no triangles
exposure(popn)

## with triangles
exposure(popn, triangles = TRUE)

## with triangles, openTriangles is "standard"
exposure(popn, triangles = TRUE, openTriangles = "standard")


## cohort
popn <- Counts(array(4:1,
                     dim = c(1, 4),
                     dimnames = list(cohort = "2001-2005",
                                     age = c(0, 5, 10, 15))))
exposure(popn)
}
\seealso{
To calculate exposure for births, see \code{\link{exposureBirths}}
To calculate exposure for origin-destination or parent-child
arrays, see \code{\link{addPair}}
}
