% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Counts-methods.R
\name{exposure}
\alias{exposure}
\alias{exposure,Counts-method}
\title{Calculate exposure from population counts.}
\usage{
exposure(object, triangles = FALSE)

\S4method{exposure}{Counts}(object, triangles = FALSE)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Counts}}.}

\item{triangles}{Logical.  If \code{TRUE}, the return value
includes a dimension with \code{\link{dimtype}} code{"triangle"}.
Defaults to \code{FALSE}.}
}
\value{
An object of class \code{\linkS4class{Counts}}.
}
\description{
\code{object} typically has a dimension with \code{\link{dimtype}}
\code{time} and \code{dimscale} \code{"Points"}, and a dimension
with dimtype \code{"age"} and dimscale \code{"Intervals"}.
The complete set of choices for age-time-cohort dimensions and
dimscales is
\tabular{ccc}{
  \code{time} \tab \code{age} \tab \code{cohort} \cr
  \code{Points} \tab \code{Intervals} \tab <none> \cr
  \code{Points} \tab <none> \tab <none> \cr
  <none> \tab \code{Points} \tab <none> \cr
  \code{Points} \tab <none> \tab \code{Intervals} \cr
  <none> \tab \code{Points} \tab \code{Intervals}
}
}
\details{
If \code{triangles} is \code{TRUE}, then exposure is calculated
separately for lower and upper Lexis triangles.  \code{triangles}
can only by \code{TRUE} if  \code{object} has dimensions with
\code{\link{dimtype}} \code{"time"} and \code{"age"}.
}
\examples{
## prepare data
library(demdata)
popn <- demdata::nz.popn.reg
popn <- Counts(popn, dimscales = c(year = "Points"))
popn <- subarray(popn, year \%in\% c(1996, 2001, 2006, 2011, 2016))
popn <- collapseDimension(popn, dimension = "region")

## no triangles
exposure(popn)

## with triangles
exposure(popn, triangles = TRUE)

## cohort
popn <- Counts(array(4:1,
                     dim = c(1, 4),
                     dimnames = list(cohort = "2001-2005",
                                     age = c(0, 5, 10, 15))))
exposure(popn)
}
\seealso{
To calculate exposure for births, see \code{\link{exposureBirths}}
To calculate exposure for origin-destination or parent-child
arrays, see \code{\link{addPair}}.
}
