% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Component-methods.R,
%   R/Counts-methods.R, R/CountsWithSubtotals-methods.R,
%   R/DemographicAccount-methods.R, R/DemographicArray-methods.R,
%   R/DimScale-methods.R, R/Intervals-methods.R, R/MetaData-methods.R,
%   R/Movements-methods.R, R/Quantiles-methods.R,
%   R/SummaryDemographic-generator.R, R/SummaryDemographic-methods.R,
%   R/Values-methods.R
\name{internal-methods}
\alias{internal-methods}
\alias{aperm,Net-method}
\alias{aperm,Pool-method}
\alias{Ops,Counts,Counts-method}
\alias{Ops,Counts,Values-method}
\alias{Ops,Counts,numeric-method}
\alias{Ops,numeric,Counts-method}
\alias{Ops,Counts,array-method}
\alias{Ops,array,Counts-method}
\alias{Ops,Counts,table-method}
\alias{Ops,table,Counts-method}
\alias{Ops,Counts,xtabs-method}
\alias{Ops,xtabs,Counts-method}
\alias{[,CountsWithSubtotals,ANY,ANY,ANY-method}
\alias{show,CountsWithSubtotals-method}
\alias{show,DemographicAccount-method}
\alias{[,DemographicArray,ANY,ANY,ANY-method}
\alias{aperm,DemographicArray-method}
\alias{as.array,DemographicArray-method}
\alias{as.matrix,DemographicArray-method}
\alias{dim,DemographicArray-method}
\alias{dim<-,DemographicArray-method}
\alias{dimnames,DemographicArray-method}
\alias{dimnames<-,DemographicArray-method}
\alias{dimscales,DemographicArray-method}
\alias{dimscales<-,DemographicArray-method}
\alias{dimtypes,DemographicArray-method}
\alias{dimtypes<-,DemographicArray-method}
\alias{drop,DemographicArray-method}
\alias{show,DemographicArray-method}
\alias{t,DemographicArray-method}
\alias{unname,DemographicArray-method}
\alias{[,DimScale,integer,ANY,ANY-method}
\alias{\%in\%,DimScale-method}
\alias{Compare,DimScale,ANY-method}
\alias{Compare,ANY,DimScale-method}
\alias{Compare,DimScale,DimScale-method}
\alias{labels,DimScale-method}
\alias{length,DimScale-method}
\alias{[,Intervals,integer,ANY,ANY-method}
\alias{\%in\%,Intervals-method}
\alias{Compare,Intervals,ANY-method}
\alias{Compare,ANY,Intervals-method}
\alias{Compare,Intervals,character-method}
\alias{Compare,character,Intervals-method}
\alias{Compare,Intervals,Intervals-method}
\alias{labels,Intervals-method}
\alias{length,Intervals-method}
\alias{[,MetaData,numeric,ANY,ANY-method}
\alias{[,MetaData,character,ANY,ANY-method}
\alias{[,MetaData,logical,ANY,ANY-method}
\alias{dim,MetaData-method}
\alias{dimnames,MetaData-method}
\alias{length,MetaData-method}
\alias{show,MetaData-method}
\alias{summary,Movements-method}
\alias{\%in\%,Quantiles-method}
\alias{Compare,Quantiles,ANY-method}
\alias{Compare,ANY,Quantiles-method}
\alias{Compare,Quantiles,Quantiles-method}
\alias{diff,Quantiles-method}
\alias{labels,Quantiles-method}
\alias{summary,DemographicArray-method}
\alias{show,SummaryDemographicArray-method}
\alias{Ops,Values,Counts-method}
\alias{Ops,Values,Values-method}
\alias{Ops,Values,numeric-method}
\alias{Ops,numeric,Values-method}
\alias{Ops,Values,array-method}
\alias{Ops,array,Values-method}
\alias{Ops,Values,table-method}
\alias{Ops,table,Values-method}
\alias{Ops,Values,xtabs-method}
\alias{Ops,xtabs,Values-method}
\title{Internal methods}
\usage{
\S4method{aperm}{Net}(a, perm, resize = TRUE, keep.class = TRUE, ...)

\S4method{aperm}{Pool}(a, perm, resize = TRUE, keep.class = TRUE, ...)

\S4method{Ops}{Counts,Counts}(e1, e2)

\S4method{Ops}{Counts,Values}(e1, e2)

\S4method{Ops}{Counts,numeric}(e1, e2)

\S4method{Ops}{numeric,Counts}(e1, e2)

\S4method{Ops}{Counts,array}(e1, e2)

\S4method{Ops}{array,Counts}(e1, e2)

\S4method{Ops}{Counts,table}(e1, e2)

\S4method{Ops}{table,Counts}(e1, e2)

\S4method{Ops}{Counts,xtabs}(e1, e2)

\S4method{Ops}{xtabs,Counts}(e1, e2)

\S4method{[}{CountsWithSubtotals,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{CountsWithSubtotals}(object)

\S4method{show}{DemographicAccount}(object)

\S4method{[}{DemographicArray,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{aperm}{DemographicArray}(a, perm, resize = TRUE, keep.class = TRUE, ...)

\S4method{as.array}{DemographicArray}(x, ...)

\S4method{as.matrix}{DemographicArray}(x, ...)

\S4method{dim}{DemographicArray}(x)

\S4method{dim}{DemographicArray}(x) <- value

\S4method{dimnames}{DemographicArray}(x)

\S4method{dimnames}{DemographicArray}(x) <- value

\S4method{dimscales}{DemographicArray}(object, use.names = TRUE)

\S4method{dimscales}{DemographicArray}(object) <- value

\S4method{dimtypes}{DemographicArray}(object, use.names = TRUE)

\S4method{dimtypes}{DemographicArray}(object) <- value

\S4method{drop}{DemographicArray}(x)

\S4method{show}{DemographicArray}(object)

\S4method{t}{DemographicArray}(x)

\S4method{unname}{DemographicArray}(obj, force = FALSE)

\S4method{[}{DimScale,integer,ANY,ANY}(x, i)

\S4method{\%in\%}{DimScale}(x, table)

\S4method{Compare}{DimScale,ANY}(e1, e2)

\S4method{Compare}{ANY,DimScale}(e1, e2)

\S4method{Compare}{DimScale,DimScale}(e1, e2)

\S4method{labels}{DimScale}(object)

\S4method{length}{DimScale}(x)

\S4method{[}{Intervals,integer,ANY,ANY}(x, i)

\S4method{\%in\%}{Intervals}(x, table)

\S4method{Compare}{Intervals,ANY}(e1, e2)

\S4method{Compare}{ANY,Intervals}(e1, e2)

\S4method{Compare}{Intervals,character}(e1, e2)

\S4method{Compare}{character,Intervals}(e1, e2)

\S4method{Compare}{Intervals,Intervals}(e1, e2)

\S4method{labels}{Intervals}(object)

\S4method{length}{Intervals}(x)

\S4method{[}{MetaData,numeric,ANY,ANY}(x, i)

\S4method{[}{MetaData,character,ANY,ANY}(x, i)

\S4method{[}{MetaData,logical,ANY,ANY}(x, i)

\S4method{dim}{MetaData}(x)

\S4method{dimnames}{MetaData}(x)

\S4method{length}{MetaData}(x)

\S4method{show}{MetaData}(object)

\S4method{summary}{Movements}(object)

\S4method{\%in\%}{Quantiles}(x, table)

\S4method{Compare}{Quantiles,ANY}(e1, e2)

\S4method{Compare}{ANY,Quantiles}(e1, e2)

\S4method{Compare}{Quantiles,Quantiles}(e1, e2)

\S4method{diff}{Quantiles}(x)

\S4method{labels}{Quantiles}(object)

\S4method{summary}{DemographicArray}(object, ...)

\S4method{show}{SummaryDemographicArray}(object)

\S4method{Ops}{Values,Counts}(e1, e2)

\S4method{Ops}{Values,Values}(e1, e2)

\S4method{Ops}{Values,numeric}(e1, e2)

\S4method{Ops}{numeric,Values}(e1, e2)

\S4method{Ops}{Values,array}(e1, e2)

\S4method{Ops}{array,Values}(e1, e2)

\S4method{Ops}{Values,table}(e1, e2)

\S4method{Ops}{table,Values}(e1, e2)

\S4method{Ops}{Values,xtabs}(e1, e2)

\S4method{Ops}{xtabs,Values}(e1, e2)
}
\arguments{
\item{a}{Typically an array.}

\item{perm}{A vector of indices or names specifying the permutation.}

\item{resize}{Logical.  Whether the object should be resized as well
as having its elements rearranged.}

\item{keep.class}{Logical.  Whether the result should has the same
class as the original.}

\item{\dots}{Other arguments.}

\item{e1}{First element in operation or comparison.}

\item{e2}{Second element in operation or comparison.}

\item{x}{Typically an object of class \code{\linkS4class{DemographicArray}}.}

\item{i}{Index.}

\item{j}{Index.}

\item{drop}{Logical. Whether to drop dimensions of length 1.}

\item{object}{Typically an object of class
\code{\linkS4class{DemographicArray}}.}

\item{value}{Replacement value.}

\item{use.names}{Logical. Whether the result should include the names
of dimensions.}

\item{obj}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{force}{A standard argument for \code{\link[base]{unname}} but
irrelevant for arrays.}

\item{table}{Object of class \code{"DimScale"}.}
}
\description{
Internal methods for existing functions
}
