% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/DemographicArray-methods.R
\docType{methods}
\name{intervalContainsTruth}
\alias{intervalContainsTruth}
\alias{intervalContainsTruth,DemographicArray,DemographicArray-method}
\alias{intervalContainsTruth,DemographicArray,numeric-method}
\title{Test whether intervals contain true values}
\usage{
intervalContainsTruth(interval, truth, lower.inclusive = TRUE,
  upper.inclusive = TRUE)


  \S4method{intervalContainsTruth}{DemographicArray,DemographicArray}(interval,
  truth, lower.inclusive = TRUE, upper.inclusive = TRUE)

\S4method{intervalContainsTruth}{DemographicArray,numeric}(interval, truth,
  lower.inclusive = TRUE, upper.inclusive = TRUE)
}
\arguments{
\item{interval}{A \code{\linkS4class{DemographicArray}},
with a quantile dimension of length 2.}

\item{truth}{A \code{\linkS4class{DemographicArray}}, or a single
number.}

\item{lower.inclusive}{Logical. Whether the lower bounds
specified in \code{interval} are considered to lie within
the intervals.}

\item{upper.inclusive}{Logical. Whether the upper bounds
specified in \code{interval} are considered to lie
within the intervals. Defaults to \code{TRUE}.}
}
\value{
An array of logical values with the same dimensions
as \code{truth}, or, if \code{truth} is a number,
a single logical value.
}
\description{
Test whether intervals contain values.  These values
are referred to as the truth, because of the way in
which \code{intervalContainsTruth} is typically used,
which is within simulations. In simulations, the
analyst generates the true rates or counts, and hence
knows what they are. \code{intervalContainsTruth} is used
to keep track of the number of intervals that contain
the true values.  This information is then used to calculate
coverage rates, ie frequency with which credible interval
of a given nominal coverage rate actually contains the
true value.
}
\details{
\code{interval} should have the same dimensions
and dimscales as \code{truth}, but also have a dimension
with length 2 and \code{\link{dimtype}}
\code{"quantile"}. The quantile dimension specifies
the upper and lower bounds. See below for an example.

The \code{lower.inclusive} and \code{upper.inclusive}
arguments can be used to specify whether the intervals are
open or closed at the boundaries. The difference between
open and closed boundaries is typically unimportant
when \code{value} consists of real numbers, but can
be important when \code{value} consists of integers, especially
when the integers are small.

\code{truth} can be a single number, in which case all dimensions
of \code{interval}, other than the quantile dimension, should
have length 1.

\code{valueInInterval} is stricter about the compatibility
of its arguments than most functions in \code{dembase}.
Although it reorders dimensions and categories
in \code{interval} and \code{truth}, it does not collapse
or expand dimensions, or drop any levels.
}
\examples{
interval <- Values(array(c(-0.1, 0.4, 0.2, 0.7),
                         dim = c(2, 2),
                         dimnames = list(sex = c("Female", "Male"),
                                         quantile = c("2.5\%", "97.5\%"))))
truth <- ValuesOne(c(0.3, 0.5),
                   labels = c("Female", "Male"),
                   name = "sex")
intervalContainsTruth(interval = interval,
                      truth = truth)

interval <- ValuesOne(c(5L, 8L),
                      labels = c("5\%", "95\%"),
                      name = "quantile")
truth <- 5L
intervalContainsTruth(interval = interval,
                      truth = truth)
intervalContainsTruth(interval = interval,
                      truth = truth,
                      lower.inclusive = FALSE)
}
\seealso{
Function \code{\link{credibleInterval}} is a convenience
function for creating \code{interval} objects of the right form
for \code{intervalContainsTruth}. Function \code{\link{MSE}},
is another function that may be useful in simulation studies.
}
