% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Component-methods.R,
%   R/DemographicArray-methods.R, R/MetaData-methods.R, R/Population-methods.R
\docType{methods}
\name{midpoints}
\alias{midpoints}
\alias{midpoints,Component,ANY-method}
\alias{midpoints,Component,missing-method}
\alias{midpoints,DemographicArray,ANY-method}
\alias{midpoints,DemographicArray,missing-method}
\alias{midpoints,MetaData,integer-method}
\alias{midpoints,Population,ANY-method}
\alias{midpoints,Population,missing-method}
\title{Replace intervals by their midpoints.}
\usage{
midpoints(object, dimension)

\S4method{midpoints}{Component,ANY}(object, dimension)

\S4method{midpoints}{Component,missing}(object)

\S4method{midpoints}{DemographicArray,ANY}(object, dimension)

\S4method{midpoints}{DemographicArray,missing}(object)

\S4method{midpoints}{MetaData,integer}(object, dimension)

\S4method{midpoints}{Population,ANY}(object, dimension)

\S4method{midpoints}{Population,missing}(object)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{DemographicArray}}.}

\item{dimension}{Names or indices of dimensions to be converted.  If
omitted, all dimensions with dimtype \code{"Intervals"} are converted.}
}
\value{
Object of class \code{\linkS4class{DemographicArray}}.
}
\description{
Change \code{"Intervals"} \code{\link{dimscales}} to \code{"Points"},
replacing each interval with its midpoint.  This can be useful for plotting
or calculating mean values, among other things. \code{midpoints} is
typically called by other functions, but it may
sometimes be useful to call it directly.
}
\details{
An object can only have a dimension with \code{\link{dimtype}}
\code{"triangle"} if the object also has age and time dimensions with
dimscale \code{"Intervals"}.  When \code{midpoints} is called
on an object with a triangles dimension, the dimtype for that
dimension is coerced to \code{"state"} and the dimscale is
coerced to \code{"Categories"}.
}
\examples{
library(demdata)
rates <- Values(VADeaths2)
midpoints(rates)

a <- array(1:4,
           dim = c(2, 2),
           dimnames = list(age = c("0-39", "40+"),
                           period = c("2001-2010", "2011-2020")))
x <- Counts(a)
midpoints(x)
midpoints(x, dimension = c("age", "period"))
midpoints(x, dimension = "period")
midpoints(x, dimension = 2)
}
\seealso{
\code{\link{dimscales}}, \code{\link{dplot}}
}
