% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R, R/Values-methods.R
\docType{methods}
\name{tfr}
\alias{tfr}
\alias{tfr,Values-method}
\title{Calculate total fertility rates.}
\usage{
tfr(object, sex = c("sex", "gender"))

\S4method{tfr}{Values}(object, sex = c("sex", "gender"))
}
\arguments{
\item{object}{An object of class \code{\linkS4class{Values}}.
\code{object} must have a dimension with \code{\link{dimtype}}
\code{"age"}.}

\item{sex}{Name of the "sex" variable, if the name is not \code{"sex"}
or \code{"gender"} (case insensitive).}
}
\value{
An object of class \code{\linkS4class{Values}}, with no
"age", "sex", or Lexis triangle dimensions, or a numeric vector.
}
\description{
Calculate total fertility rates from age-specific fertility rates.
}
\details{
The total fertility rate (TFR) is a summary measure commonly used by
demographers.  It is the number of births that the average woman
have if she survived to the end the reproductive ages, and if
current age-specific fertility rates were to persist indefinitely.

\code{object} will typically contain dimensions other than age,
such as time or region.  Separate total fertility rates are calculated
for each combination of these other variables.

The calculations are done aggregating the "age", and, if present,
"sex" and Lexis triangle dimensions.  See \code{\link{dimtypes}}
for more on age and Lexis triangle dimtypes.) If \code{object} contains
a "sex" dimension, this is assumed to be the sex of the child,
not the parent. Total fertility rates do not distinguish
female and male births.

A dimension is assumed to measure sex if it has the name
\code{"sex"} or \code{"gender"} (case insenstive).  If another
name is used, function \code{tfr} can be informed via the \code{sex}
argument.

If \code{object} has iterations, capturing uncertainty, so will
the return value.
}
\examples{
births <- demdata::nz.births
popn <- demdata::nz.popn.reg
births <- Counts(births, dimscales = c(year = "Intervals"))
## use mid-year population to approximate person-years lived
## over year
popn <- Counts(popn, dimscales = c(year = "Intervals"))
females <- subarray(popn, sex == "Female")
fert.rates <- births / females
tfr(fert.rates)

## use of 'sex' argument
fake.data <- array(runif(4, max = 0.05),
                   dim = c(2, 2),
                   dimnames = list(age = c("15-29", "30-49"),
                       "sex of child" = c("Female", "Male")))
fake.data <- Values(fake.data)
tfr(fake.data, sex = "sex of child")
}

