% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous-functions.R
\name{yearToPeriod}
\alias{yearToPeriod}
\title{Convert years to periods}
\usage{
yearToPeriod(year, breaks, firstOpen = FALSE, lastOpen = FALSE)
}
\arguments{
\item{year}{A vector of years. A numeric vector, or a vector
than can be coerced to numeric.}

\item{breaks}{A vector of breaks, specifying the points dividing
periods.}

\item{firstOpen}{Logical. Whether the first period is "open",
i.e. has no lower bound. Defaults to \code{FALSE}.}

\item{lastOpen}{Logical. Whether the last period is "open",
i.e. has no upper bound. Defaults to \code{FALSE}.}
}
\value{
A factor, the same length as \code{year}.
}
\description{
Convert a vector of years to a vector of periods.
The periods are formatted in the way expected by
functions such as \code{\link{Counts}} and
\code{\link{Values}}.
}
\details{
The years can be contain decimal fractions such as \code{2000.25} or
\code{2018.633}, but are more typically integers
such as \code{2000} or \code{2018}.

If \code{year} is a factor, then \code{yearToPeriod} will coerce
it to a character vector before trying to coerce it to numeric.
See below for an example.
}
\examples{
year <- c(2001, 2023, 2000, 2005, 2014, 2013, 2029)
## 5-year periods, 2000-2005, 2005-2010, ..., 2025-2030
yearToPeriod(year, breaks = seq(2000, 2030, 5))
## 5-year periods, 2000-2005, 2005-2010, ..., 2045-2050
yearToPeriod(year, breaks = seq(2000, 2050, 5))
## first period open 
yearToPeriod(year, breaks = seq(2010, 2030, 5), firstOpen = TRUE)
}
\seealso{
\code{\link{ageToAgeGroup}}
}
